Fix: USD export: Prim name collision
Fixed name collision when processing names with non-ascii characters.
Now taking the names of existing objects and data in the scene into
account when computing new prim names.

Also, for now computing custom names and saving display names only
for names that contain non-ascii characters.  Previously, this was
done for any names that are invalid for USD.  This behavior was
reverted for now to retain more information in the original name.
We can revisit handling display names for invalid USD names in the
future.