Fix: GPv3: Assert in `LayerTreeView::build_tree()`
This was caused by 745fd2a2cb58e40337310c28ab1cf4a418335d0c.
The issue was that there was an attempt at calling
`uncollapse_by_default` on a `LayerViewItem`
which can't be collapsed (because it can't have any children).
This then triggered the assert.

The fix removes the call to `uncollapse_by_default` for
`LayerViewItem`.