Brush Assets: Improve defaults for "Duplicate Asset" operator
Use the library from the operator's last execution instead of storing a
"default" option in the user preferences. Or if the last library isn't available,
use the current library as a default, or the first library if the current
library isn't editable.

Also use the asset's current catalog path for the initial asset path.

Pull Request: https://projects.blender.org/blender/blender/pulls/120028