Fix #119973: Flip node produces artifacts
The Flip node produces artifacts if not connected directly to an input.
This due to a wrong calculation of depending area of interest, which is
a regressed introduced in 70a8a9e.

This is fixed by reverting the removal of the area of interest
calculation method in the aforementioned commit.

Pull Request: https://projects.blender.org/blender/blender/pulls/120149