Fix: Sanitize filename for collection export
When creating a filename for use with a File Handler, we should guard
against problematic characters like "/" and "\\", among others, which are
not safe to use in paths. Sanitize the incoming name with
`BLI_path_make_safe_filename` to ensure the name can be used.

For empty incoming names, or names containing all spaces, we default to
"untitled" before adding the extension.

Pull Request: https://projects.blender.org/blender/blender/pulls/120652