Vulkan: Missing include when compiling draw tests
MoltenVK includes are not found when compiling draw tests on certain
configurations. It has to do with the number of ways the VulkanSDK and
MoltenVK can be configured. Other executables the molten vk was included
to the system includes, except when compiling drawing tests.

This PR will equalize the includes how MoltenVK between the different
binaries

Pull Request: https://projects.blender.org/blender/blender/pulls/120888