GPv3: Tint tool UI fixes
This PR contains 3 fixes in the tool bar for the Tint tool:
- For the tint color, the general color panel was shown, but it should
be the Grease Pencil color panel (with not just the color wheel, but
also the color palettes).
- The Advanced Brush panel was redundant, because it only showed
Radius and Strength and those are already present in the tool bar.
So the Advanced panel can be removed.
- The 'Active Layer' switch wasn't exposed in the tool bar, so it has
been added.

Pull Request: https://projects.blender.org/blender/blender/pulls/120719