Fix: Issues with asset shelf in image editor
Make the asset shelf always show in the image editor's paint
UI mode. And fix the popup to show in the sidebar. That required
changing the way we retrieve the brush asset shelf name from the
context. ALso fix a few panels that didn't check for whether there
was a brush. While that shouldn't really happen it does currently
and it's nice to avoid the errors.