Brush Assets: Add operator to choose brush preview image
A simple operator like the existing `ed.lib_id_load_custom_preview`
operator. I didn't reuse that because the requirements are a bit
different. In particular we want to re-save the asset when the
preview image is changed.

The operator is available as a new entry in the existing brush asset
context menu.

Pull Request: https://projects.blender.org/blender/blender/pulls/121505