Geometry Nodes: support matrix in Transform Geometry node
Now the Transform Geometry node either has individual components (location,
rotation and scale) or a matrix as input. In many cases that is just more
convenient when working with matrices. Often this is identical to splitting
the matrix into it's individual components. However, if the matrix contains
shearing, information is lost when splitting it into individual components.

Pull Request: https://projects.blender.org/blender/blender/pulls/121438