Brush Assets: Implement grease pencil default eraser brush
Instead of a boolean on brushes, this is now a secondary brush pointer
and asset reference. It can be chosen from a new Eraser panel in the
brush tool properties.

It was implemented generically, since this functionality may be useful
to extend to other modes in the future.

This currently defaults to Eraser Hard, should be changed to Eraser
Soft once the essentials library has been updated to include this.

Note: actually selecting the brush depends on #121597.

Pull Request: https://projects.blender.org/blender/blender/pulls/121675