Fix #115460: EEVEE-Next: Missing bind when drawing curves
When drawing hair/curves the material shader has a "a" texture slot,
but this slot was never bound to a texture. This PR fixes this by
adding a default texture to it and when available tries to bind
CD_AUTO_FROM_NAME resources.

This part seems to be not implemented when the new curve rendering
was added. `CD_AUTO_FROM_NAME` was never considered and therefore the
layer could not be found and the sampler could not be found.

Fixes #115460

Pull Request: https://projects.blender.org/blender/blender/pulls/121064