Brush assets: Use Popover for asset shelves & operator to invoke from shortcut
- Turn the asset shelf popup into a popover. This requires instantiating it via
  a panel.
- Add operator to call an asset shelf popover. It takes the idname of the asset
  shelf to display as operator property.
- Allow assigning a shortcut for an asset shelf popover through the context
  menu of buttons invoking the asset shelf popover ("Assign Shortcut", "Remove
  Shortcut" and "Change Shortcut").
- Add Shift+spacebar as asset panel popover shortcut for testing. Unclear if we
  want to keep it there.
- Removed change to always support refreshing popup blocks flagged with
  `UI_BLOCK_KEEP_OPEN`. Might make sense still, but will keep it separate.

Had to add support for storing strings as context members, to pass the asset
shelf name to the panel code somehow before the panel is created even. Think
this is generally useful, as we've tried to pass such strings through context
before. We'd fallback to creating RNA structs holding a single string property
to be able to do this, an annoying workaround.

Pull Request: https://projects.blender.org/blender/blender/pulls/122068