EEVEE-Next: Improve shadow map filter
This make the filter have a constant width in shadow
space making sure to always filter the correct amount
of pixels.

This uses the receiver slope to put the samples on a
cone with its apex at the shading position instead of
setting up the disk in shading tangent space.

We limit the slope bias to 45 degrees and offset the
cone in normal direction if the angle between light and
receiver is greater than 45 degrees. This avoid any
self shadowing artifacts cause by this technique.

Pull Request: https://projects.blender.org/blender/blender/pulls/122266