GPv3: Draw Tool: Randomize radius and opacity
This adds the randomization options for radius and opacity (strength).

Instead of pure white noise (like in GPv2) this uses perlin noise to make
the effect less jagged. The scale of the noise (in screen space) is
hardcoded to roughly fit GPv2. It could be exposed as a setting in the
future to get noise that changes slower/faster along the stroke.

Pull Request: https://projects.blender.org/blender/blender/pulls/123715