Sculpt: Data oriented refactor for clay thumb brush
Part of #118145.
Since the initial commit (015d5eda884dfb10f898) there has been a logic
mistake in the code for this brush where a variable was unused. The code
has worked for years in that state, so I left it in. The next commit
will clean up the unused code.