Rename "Sculpt Tool" etc to "Brush Type" in the UI
The name "tool" is overloaded, plus, there actually is a "Brush Tool" now. As a
follow up to the assets project merge we can rename this in C++, added to #116337
and opened #124201. At least rename it in the UI for a start.

Pull Request: https://projects.blender.org/blender/blender/pulls/124203