UI: Allow asset shelf to not have an active item
For brush assets we want to display brush assets with any tool selected, but
only highlight the active brush asset when the "Brush" tool is enabled. It can
be confusing to have the brush highlighted when it doesn't actually do anything
currently.

Initially reviewed in:
https://projects.blender.org/blender/blender/pulls/121671

Part of the brush assets project, see:
- https://projects.blender.org/blender/blender/issues/116337
- https://projects.blender.org/blender/blender/pulls/106303