Fix #124342: Color sampling clamped to 1 for specific tool
SHift-X sets the color value directly, bypassing clamping done
on the RNA level. The color picker uses RNA setter to assign
colors, which enforces limits.

This change makes it so the 0..1 is an UI range, but allows to
manually type in values above that. This is similar to colors
in the node trees, but a bit more conservative by using UI range
explicitly.

Thanks Omar for identifying the root cause of where the clamping
is coming from!

Pull Request: https://projects.blender.org/blender/blender/pulls/124355