Fix: More GPU subdivision crashes with loose edges
The previous fix stopped creating these vertex buffers when there
is no faces, but that introduced the need for a bunch of null checks
around the extraction code. But the important part is not binding
the empty VBOs.

Also it seems there was a merge error in the previous fix,
that's corrected here too.