Fix #126607: GPv3 SVG import has wrong object transform and missing gradient color
The imported GP object should start out in default X/Z plane
orientation. The `EulerXYZ` function expects radians.

Gradient colors are not supported by Grease Pencil. GPv2 was simple
ignoring them and using the `NSVGgradient` pointer as a "color" value,
leading to effectively random colors. GPv3 was outputting black instead,
but that is easy to mistake for a deliberate color. This patch computes
an average gradient color as an approximation, and also generates a
warning for the operator when gradients are encountered.

Using magenta as a visual indicator is also a possibility, but this
proxy color seems to work pretty well.

Pull Request: https://projects.blender.org/blender/blender/pulls/126650