Fix #126137: Unable to create pose asset with layered action
The issue was that the `id_root` property can be 0,
but the enum doesn't officially support this value.

The fix is to create a new function `rna_id_root_itemf`
that dynamically creates the enums, which is basically
`rna_enum_id_type_items` + an extra entry for 0.

Pull Request: https://projects.blender.org/blender/blender/pulls/126837