Fix #126986: No background when only rendering volumes in 4.2.1
This is a fix that is only happening in 4.2.1. In 4.2.1 a fix was added
that changed the way how backgrounds where integrated. This assumed that
the combined framebuffer was bound. In 4.2.1 this is not always the
case.

> NOTE: this fix will only be applied to 4.2.1 as 4.3 doesn't have this issue.

Q: Should we only do this when combined_fb is not bound?
Pull Request: https://projects.blender.org/blender/blender/pulls/127068