Build: Add HIP-RT libraries to dependencies builder
This change makes it so that HIP-RT libraries are built as a part
of the rest of the dependencies.

The libraries are built and placed to the `lib-<platform>/hiprt`. They
are not currently used, but are needed to simplify transition of Cycles
to these new libraries.

The actual Cycles side is done in #121050

Pull Request: https://projects.blender.org/blender/blender/pulls/127719