Vulkan: Enable in a regular release
The Vulkan backend has come a long way and it is time to spread his
wings a bit more. This change will make the vulkan backend available
in a regular release starting from 4.3.

Vulkan is still experimental and performance isn't there yet. But it
gives add-on developers already time to check if their add-ons work.
Main goal for the vulkan project is to get a better understanding which
platforms are supported.

- NVIDIA GTX900 and up should be supported using the official NVIDIA
  drivers. This means that support for GTX700/800 is currently not
  available. This is related to dynamic rendering.
- AMD Polaris and up are supported using the AMD drivers. When using
  the open source drivers I believe the support level is around GCN2.
- Intel CPUs are supported from UHD (6th gen) on Linux. On windows
  it requires features that are not available on all latest drivers and
  a workaround should be added on our side.
- Intel GPUs are supported preferrable using the latest drivers on
  Windows and Linux

Some render artifacts are showing when using EEVEE but the cause is
clear and would assume to solve them before the actual release.

Pull Request: https://projects.blender.org/blender/blender/pulls/127096