Geometry Nodes: add Iteration socket to repeat zone
For most cases where the iteration index is needed, a For Each zone is the
better than using a Repeat zone. It's easier to use and more optimizable.
However, in some cases a Repeat zone is actually the right solution and the
index often useful in such cases too.

Pull Request: https://projects.blender.org/blender/blender/pulls/127521