Spreadsheet: use more obvious fallback name for geometries
Instead of `Geometry`, the fallback name is now `(Geometry)`. The goal is to
make it more obvious that this is not the actual name, but just a fallback. With
some luck it also motivates people to use better custom names.

The same thing is done for grease pencil layer names.

Pull Request: https://projects.blender.org/blender/blender/pulls/128207