Fix #129066: ParticleSettings.material_slot wrong from python
`rna_Particle_Material_itemf` was using the context object for getting
the materials.
This is OK when doing it through the UI (the context object is correct
in the Properties Editor, also when pinned), but might not be correct
when doing this from python.

So now do similar as in `rna_Particle_change_type`, iterate all objects
to find the one matching the `ParticleSettings` (but only do this if the
context object does not have a matching `ParticleSettings` already).

NOTE: `rna_Particle_change_type` could be sped up as well with this
method (checking context object first), might be for a separate PR.

Pull Request: https://projects.blender.org/blender/blender/pulls/129101