Fix #129183: Blender freezes permanently when pausing animation with multiframe on
This was caused by multiedit state not being recorded
inside the batch cache. The stale batch from animation
playback was being reused for drawing with multiframe
range after stoping playback. This lead to out of bound
read.

The fix is to store the multiedit state inside the batch
and invalidate the batch if it changes.

This should only be applied to GPV2 releases (<=4.2).

Pull Request: https://projects.blender.org/blender/blender/pulls/129911