Fix #129719: Compositor produces black on Intel GPUs
The compositor produces black when rendering on Intel GPUs. That's
because shader output images have different precision that that of the
output textures where the output is written. To fix this, make sure a
shader of the appropriate precision is requested.

Pull Request: https://projects.blender.org/blender/blender/pulls/129969