NPR: Shader Node Traits
The NPR branch needs a way to tell whether the node tree is a group
or an NPR tree for correct polling in the UI.
This follows a similar approach to `GeometryNodeAssetTraits`.

There's still the problem that any node group can still be added to
any shader tree type (for example, it's possible to add a Material
node group to a World node tree).
This trait system could eventually be used to detect which node
groups have tree-type-specific nodes and allow adding them only to
compatible trees.

Pull Request: https://projects.blender.org/blender/blender/pulls/130192