Support custom remote asset library type, set up in Preferences
Similar to the extensions UI, clicking on the '+' icon to add an asset
library now gives you the choice to add a local, or a remote library.
When choosing a remote one, a URL can be entered. The UI in the
preferences shows only the URL field for remote libraries too.