Fix #130556: applying keyingsets in NLA stack doesn't properly inverse solve
When applying keyingsets, the `PointerRNA` for the `ID` (in this case an object) is passed in.
The function `insert_keyframes` then passed that along to `BKE_animsys_nla_remap_keyframe_values`.
with the resolved property pointer which would mean the remap fails because they are not matching
in the case of bones.
That would only apply to keyingsets, because the other code path would pass in the `PointerRNA` for the bone
already, meaning the result of resolving it is the same.

The fix is to pass the resolved `PointerRNA`

Pull Request: https://projects.blender.org/blender/blender/pulls/130679