deps_builder: Fix HIP detection on first run on windows.
The HIP compiler version was not being detected on first run in a fresh
build folder since HIP requires perl, and perl wasn't installed yet due
to the compiler detection running before the installs into the build
folder.

this change re-orders the compiler detection and environment preparation
so perl is available when we detect the hip compiler version.