Fix #130503: Separate Geometry node outputs empty grease pencil layers
I'm not quite sure what was wrong about the old implementation, but it was confusing
because it also depended on the user counts of drawings and other more high level
functions like `duplicate_layer`.

This fix just reimplements the functionality using the grease pencil API that we use in
more places in geometry nodes.

Pull Request: https://projects.blender.org/blender/blender/pulls/131291