Fix #130061: Pose Sliding with no bones selected resets overlay flags
The issue was that the struct `tPoseSlideOp` tracks the
viewport overlay flags so they can be restored after the operator has run.
But the flag was stored at the end of `pose_slide_invoke_common` which
exits early if no keys are selected.
This then calls `pose_slide_exit` which sets the viewport flags to 0, because they were
never stored in the first place.

The fix is to move the storing of the flag up to `pose_slide_init` to ensure that it
is captured as soon as the struct is allocated.

Pull Request: https://projects.blender.org/blender/blender/pulls/131327

Pull Request: https://projects.blender.org/blender/blender/pulls/131671