Fix #129416: Wrong Bone Selection Set used for selection
The issue was that when using a specific selection set,
the property `selection_set_index` is set to a specific index.
That index would persist to the next call of the operator
when trying to call it from the selection set list. The fix is to explicitly reset the
index to -1 when calling the operator from the list, because
that means the active selection set will be used.

Pull Request: https://projects.blender.org/blender/blender/pulls/131675