Fix #131669: Camera not autokeyed in special case
The autokeying code for cameras used the keyingset code to insert keys.
In the case of "Only Insert Available" turned on this would use the "Available" keyingset.

However, in the case of looking through the camera and moving the viewport
when the camera is not active, the poll function of that keyingset would return false.
Instead of modifying the poll function, the fix is to use the more direct keying code
using `RNAPath`.

This can be backported to 4.2 but not 3.6 due to the changes to the keying code done in 4.0

Pull Request: https://projects.blender.org/blender/blender/pulls/131796