Release tools: Add script to generate a list of bugs fixed from last release
This commit adds a script that will collect a list of commits that
fix bugs that have been in Blender for 1 or more major release.

This list can then be added to the release notes, an example being:
https://developer.blender.org/docs/release_notes/4.3/bugfixes/

Details about how it works can be found at the top of the script, or
by running the script with `-h`.

Pull Request: https://projects.blender.org/blender/blender/pulls/131615