Fix #134011: Overlay: Loose wires object color in selection not visible
The changes from 29356e2bf76682cbf1d33ffcd0691cf1cd707213
fixed a readability issue with loose wire not being as prominent as
regular surface wireframe.

But also added a regression where the object color wireframe mode
made the object color imperceivable when the object is selected.

This commit fixes the issue by setting the facing factor to 0.5 for
loose wires. This way, both the theme color and object color wireframe
more are blended enough to be distinguishable and still retain the
object color hint.