Fix #132363: Gaussian Blur ignores connected input
If a transform node is linked to the blur node, then the node might ignore the input connected to the `size` input socket, which results in the wrong `size` value being used.

The reason is the Blur Operation explicitly checks whether the input operation is a constant operation or not. However, the node tree builder might add some non-constant operations, e.g. corrective canvases _after_ constant folding. This breaks the assumption that a socket linked to a constant input (such as a Value Node) will remain connected to a constant input after node tree building is done.

Pull Request: https://projects.blender.org/blender/blender/pulls/133375