Fix #134770: Multires persistent base doesn't work randomly
The previous attempt to restore this behavior was in
0a2d5d58013e579bf06c67d50b49497ce23ccbcb

That commit introduced clearing the related `SculptSession` variables
when the paint BVH was freed to avoid bad behaviors seen in 4.2 and
prior. However, this solution is somewhat incorrect, as the prior state
of this data is not necessarily restored when the BVH is recreated,
unlike other temporary mapping data.

This results in the persistent base data being cleared occasionally with
no indication to the user that this is happening, causing the setting to
appear to be inactive.

To fix this, this commit makes a few changes:
* Removes clearing this data in the `BKE_sculptsession_free_pbvh`
  function.
* Initializes the displacement arrays to the same size as the position
  and normal array.
* Introduce new variables to track the saved multires grid size so
  that it is not deleted, only considered invalid when the topology
  changes.

Pull Request: https://projects.blender.org/blender/blender/pulls/134780