Fix #134909: Crazyspace has incorrect deformed positions due to armature modifier
The fix 8af27ce27ef25ae4a15c5f4c22dcce048f2fa06e assumed that
the `edit_hints->positions()` were always empty and that we would always
fallback to using the evalauted positions instead.

But if the edit hints are populated, this fails. In this case, we populate the
positions in `remember_deformed_positions_if_necessary`.

The fix makes sure that we write to the `edit_hint` positions if they are
populated.

Pull Request: https://projects.blender.org/blender/blender/pulls/135033