Tests: Add a Cycles test for #135200
A recent bug fix to light tree building lead to a issue where some
lights with light linking would incorrectly share nodes in the
light tree.

To help detect failures like this in the future, this commit adds a
test based on the file found in the original report
blender/blender#135200.

Ref: blender/blender-test-data!63