Cleanup: Remove `#ifdef` blocks for Dyntopo edge code
At this point, it is unlikely that we will revert the changes made over
a decade ago to the Dyntopo edge splitting code. The old code will still
exist in the git history, but to improve readability, this commit
flattens out the blocks and removes the `#define` statements.

Pull Request: https://projects.blender.org/blender/blender/pulls/135410