macOS: Disable OpenMP when using Xcode 16.3 / LLVM 17
The precompiled library uses incompatible ABI with this version,
leading to linking errors.

There is no easy short-term solution: as per documentation the
OpenMP library must match LLVM version. It is hard to achieve in
a way that both older and newer LLVM versions are supported: it
would need to be multiple OpenMP libraries pre-compiled and chosen
at the build time.

There are not that many places left in Blender when OpenMP is
still used, so it seems better to invest in getting of those
remaining cases. The most challenging one would be Eigen, but
need to check whether the code-paths that are used from Eigen
even contain OpenMP.

Pull Request: https://projects.blender.org/blender/blender/pulls/136828