Modeling: Add a new boolean solver based the the Manifold library.
Adds the 'manifold' solver option to the Boolean geo node and to
the Boolean modifier. This solver is about as fast, or faster,
than the current float solver, and is robust against floating
point issues like the Exact solver.  But currently it only
works on mesh arguments that are strictly manifold.

See https://projects.blender.org/blender/blender/issues/120182
for many more details.

This depends on the Manifold library being added as a dependency,
see https://projects.blender.org/blender/blender/pulls/135261 .
This won't compile until some tweaks are made to reflect that.