Fix: Retain slot name when baking actions
Previously when an action was baked, the slot name was not retained.
This causes problems when switching between actions because the slot
will not automatically be assigned.

This is now fixed by ensuring that the name of the last assigned slot
is used to create the new slot.

----

Backported from [#136814: Fix: Retain slot name when baking action](https://projects.blender.org/blender/blender/pulls/136814)
Change is exactly the same, except for the unit test code because the test in question doesn't exist for 4.4

Pull Request: https://projects.blender.org/blender/blender/pulls/136925