Fix: VSE: Local transform for mirrored strip images
When creating the `axismtx` in `SeqToTransData`, the strip's rotation
needs to be multiplied by the mirror values.

This adjustment effectively multiplies the rotation by -1 when a mirror
is active, resulting in the `axismtx` being correct. This process is
essentially the same as in #136004.

The `axismtx` is utilized in various transform modes, such as
`transform_mode_translate`.

Pull Request: https://projects.blender.org/blender/blender/pulls/137489