Fix: EEVEE: Missing shadow update during light baking
This is the root cause of the non-deterministic behavior of
light baking. The shadow system was not given enough update
iteration to fully render all tiles of the Virtual shadow maps.

This would resulte in missing tiles inside the bakes.
The value of the tiles seems to have been implementation
dependant, which, on some implementation created light leaks.

This fixes some recent regressions on the MacOS builbot tests.

This is candidate for backporting to 4.2.

Fix #138147

Pull Request: https://projects.blender.org/blender/blender/pulls/138203