Hydra: Add selection of Vulkan HGI
When using Vulkan HgiOpenGL was still selected. This PR changes the Hgi selectiong
to be based on the active GPU backend.

With this PR the correct Hgi could be selected, this requires the USD/Hydra libs to be
compiled with Vulkan support.

The previous implementation used `HGI_ENABLE_VULKAN` but had an issue:
- Depending on the platform the environment variable was stored inside a global
  and initialized when loading blender (even before our main entry point is triggered)
- Storm render engine in Blender has been adapted to do not do any detection.

Ref #133717

Pull Request: https://projects.blender.org/blender/blender/pulls/135780