Spreadsheet: reuse same table for Viewer nodes in different loop iterations
Previously, each Viewer node in a different loop iteration would have a separate
table with a separate column ordering. This is annoying when using the
inspection index to look at different iterations.

This patch changes it so that the iteration index of the table identifier is
ignored.

Pull Request: https://projects.blender.org/blender/blender/pulls/139511