Fix: USD: Set midlevel and scale inputs for constant displacement
Explicitly set the Midlevel and Scale values for Displacement material
nodes during Import. This is necessary for "constant" displacement,
where the incoming UsdPreviewSurface does not have midlevel or scale
data associated with any part of the node graph. We default to Midlevel
and Scale values of 0.5 and 1.0 respectively; as that's what the prior
code assumed, and will closely match what is seen in `usdview`.

Pull Request: https://projects.blender.org/blender/blender/pulls/139510