Fix #139157: Fix Movement of Numbers in animation editors
Both horizontal and vertical scale indicators rely on string measuring
for proper positioning. But in order to always do this reliably the
font's size needs to be set explicitly. Otherwise the font size could
be incorrect from prior use so the measurement is wrong, but then the
size is correctly set when drawing occurs. It draws at the right size
but at wrong position.

Pull Request: https://projects.blender.org/blender/blender/pulls/139549