Fix #137735: Scale Icon Border Intensity by Icon Alpha
When an icon is drawn with both lowered opacity and a border shadow
(used to increased contrast in Light theme), it will look darker
because more of the dark shadow is visible through the icon. This PR
fixes this by scaling the shadow by the opacity.

Pull Request: https://projects.blender.org/blender/blender/pulls/139610