Fix: Geometry Nodes: Show menu socket name in boolean Switch nodes
The socket name on Menu sockets is hidden by default to save layout space.
In some nodes with multiple menu inputs (switch nodes) this leads to ambiguity,
so the socket name is enabled explicitly for these node types. This was already
done for Menu Switch nodes but not for the boolean Switch node yet.

Pull Request: https://projects.blender.org/blender/blender/pulls/139628