Cycles: Use HIP SDK 6.4.0 on Linux
It seems to be fixing an issue rendering the junkshop file on RDNA4 GPUs.
Tested on Ubuntu 22.04 and 9070 XT GPU.

The downside of this change is the curve intersection in HIP-RT being
slower (figures are a bit scattered, but average to ~10%). This seems
to be caused by compiler resolving function inlines and spilling
differently. Quite unfortunate, but it is better to solve known crashes.
There might still be time to look into the performance topic during the
Beta period.

Ref #136143

Pull Request: https://projects.blender.org/blender/blender/pulls/138981