UI: use automatic search box width in more places
Most search-boxes still used a hard-coded width. This has the problem that it is
often too wide or too small for its content. This patch builds on top of
bf696320f619bf6d056ade2f236cd44098753965 to make the size of more menus adapt
automatically.

In all the cases I've tried, the width-detection code only ran once when the
search is triggered, and not continuously as might happen for some menus as
mentioned in
https://projects.blender.org/blender/blender/pulls/138653#issuecomment-1572027.

Pull Request: https://projects.blender.org/blender/blender/pulls/139423