Fix #139274: Grease Pencil: Update vertex color overlay value
In Geometry Nodes workspace, the viewport has a default value of 0.0
for `gpencil_vertex_paint_opacity`, this causes material preview to not
show proper vertex color even when strokes have color. Considering this
property is a bit obscure, setting a default value of 1.0 here makes
sense and it's also consistent with the rest of the viewport editors in
other workspaces.

Pull Request: https://projects.blender.org/blender/blender/pulls/139356