Sculpt: Remove dyntopo BMesh reordering
In certain cases, when a Mesh is generated from a BMesh as part of
Dyntopo sculpting, (e.g. when the data is invalidated when leaving
Dyntopo or Sculpt mode) the face and vert indices are compressed to
remove any gaps that may exist due to deletion of elements. This step is
redundant to perform prior to converting to a Mesh, as the process of
doing so already compresses the indices

To avoid doing extra work, this commit simply removes this conversion
and any associated helper methods.

Pull Request: https://projects.blender.org/blender/blender/pulls/139313