Refactor: move HTTP downloader metadata handling to separate class
Move the handling of the HTTP metadata (i.e. the juggling of headers
necessary for conditional downloading) to a class of its own. This way
there is a clear interface for it, and it can later be swapped out for
a similar interface using sqlite.

Also it should be easier to download to one file path, but verify the
local metadata with another file path (i.e. just-downloaded vs. validated
versions of the same file).

No functional changes.