HTTP downloader: add callback to report errors in on-done callbacks
Add a callback `on_callback_error` to the background downloader. This will
get called whenever an `on_download_done` callback raises an exception.

The `on_download_done` callback is the "motor" that keeps a complex download
process going. Basically its purpose is to parse the downloaded file,
and queue up more downloads if necessary (for example, downloading the next
page of the asset listing). If that function itself raises an exception,
the new `on_callback_error` callback will get called, so that the user
of the background downloader knows about this issue.

This is done via a callback, instead of bubbling up the exception, as that
gives a clean API for error reporting. Bubbling the exception would make
it appear when `BackgroundDownloader.update()` is called, and then it is
hard to provide it with concrete metadata about the HTTP request and the
file path it was saved to.