BLO: enable new file format by default
This enables the new .blend file format as described in #129309 and implemented
in 6a903823904628306d3818187350b550a6753c0c by default. This allows storing
blocks larger then 2GB inside the file. This is a necessary requirement to be
able to store e.g. meshes with more than a few hundred million vertices.

There is a new debug option that can be used to force Blender to write the older
file format. This is especially useful when using tools that have not been
updated to be able to read the new format yet.

Compatibility:
* Blender 4.5 is able to read the new format for a few months already, so 4.5
  LTS will be able to read these files
  (6a903823904628306d3818187350b550a6753c0c).
* The internal `blendfile.py` which is used by `blend2json.py` has been updated
  (d83bfee34718d6ed4e0e6bfb6028833a55b429bd).
* blender-asset-tracer (BAT) has been updated
  (blender/blender-asset-tracer@f1ee7980b25bfa572ee993648e997ebfb62982c2).
* `blend_render_info.py` will be updated soon (#140341).

Pull Request: https://projects.blender.org/blender/blender/pulls/140343