Remote Listing Downloader: download all thumbnails
When downloading the remote listing, also download all thumbnails of all
assets.

The thumbs are downloaded to a path that is determined by hashing the
URL of the asset itself, its ID type, and its name. This makes it
possible to find the thumbnail for a specific asset later, without
having to remember each asset's thumbnail URL.

Limitation: the thumbnail URL is expected to end in the file extension.
This makes it possible to determine the correct filename (including
extension) before the download starts, considerably simplifying the
download process.

For now only PNG and WebP URLs are accepted. Blender's thumbnail preview
system does not support WebP yet, but that limitation should be lifted
soon-ish.

The thumbnails are stored in a local directory (`$LOCAL_ROOT/_thumbs`).
They could be downloaded directly to the `$XDG_CACHE_HOME/thumbnails/`
directory, but that would require already knowing where the asset itself
will be downloaded. That will be done soon too.

There are quite a few TODOs added in this commit, to indicate
near-future work.