Remote Listing Downloader: warn when thumbnails are not images
When downloading thumbnails from a Github repository, they can be served
as text files, to be used by the Git-LFS system. This is now detected
and warned about.

The files are still kept, because there is no way yet to store "failed"
thumbnails in a way that prevents very frequent re-downloading.