Remote Asset Listing: separate 'files' from 'assets'
Separate the per-file information from the per-asset information. This
removes duplication (when there's multiple assets per blend file), and
clarifies what entity (the blend file or the assets within) the information
is about.

Also this adds a `path` property to files, which indicates where in the
asset library the file is located. This is necessary to make downloads
independent of the URL of the file (which was, until now, used to determine
this location).