Fix: UI: Extend select for tree-view not always working
Noticed that when clicking a tree-view item with multi-select support
(shape keys in this case) twice without ctrl held, and then another one
with ctrl held, the previous one would get deselected.

That's because the second click would first deselect all items and then
call `AbstractViewItem::activate()` on the clicked item. This would fail
to re-select the item, because it skipped touching selection when the
active state didn't change (as was the case here).