Fix #136359: Add data-block selector to handle missing texture better
Instead of only showing a button to switch to the texture properties -
which don't actually allow creating a texture for linestyle - add a
data-block selector that allows creating a texture for linestyle. It
will also show a button to switch to the texture properties, if
available.

It seems like the previous UI didn't have any good way of assigning a
texture for linestyle, without going into the texture node editor.

Pull Request: https://projects.blender.org/blender/blender/pulls/141408