Avoid endless preview loading when building without WebP support
When the online assets used WebP and Blender was built without, it would just
keep loading. Some timeout could help, but we can also just tag the preview as
invalid if there's a preview in a generally supported, but not readable format.
Will also tag previews as invalid if the image file can't be read for other
reasons.