Revert "Deps_builder: pydantic 2.11.7"
This reverts commit 81e862140fa24962ec9edd7b27f25147201ca4b5

binary packages turned out to have issues for debug builds on windows
as they link to python_311d.dll and none of the binary packages offfered
on pypi offered a debug build. Building pydantic from source was rejected
due to no rust build infrastructure being in place yet and deemed not
worth the effort for this one library dependency, an alternate solution
was chosen by the assets team that requested this dependency.