Fix #142163: Instance attributes can be invalid on curves
The "curve_type" attribute in curves geometry is built-in and only valid with
a `int8` type on the `Curves` attribute domain. Adding it with a different type
on instance geometry is fine though, but causes invalid attribute writer access
when realizing the instances.

Pull Request: https://projects.blender.org/blender/blender/pulls/142218