Fix #142630: Custom socket polling fails due to missing a "base" type
Adding a custom node socket in a node group fails, because the socket
type no longer passes the poll function.

Socket type polling for generated enum items was changed in #138824:
The socket type enum is based on the position in the socket types list,
but regular sockets generate many subtypes that should not appear in the
enum. Therefore only the "base type" of the socket is used for the enum
entry. Custom sockets, however, do not have a base type and are used
directly. This broke because now any socket type without a base type
is always rejected.

Fixed by handling a missing base type as indicator that the socket type
should be used directly.

Pull Request: https://projects.blender.org/blender/blender/pulls/142816