Fix #142325: Grease Pencil: Join Selection removes wrong points
The Join Selection operator can remove the original selected points, but
this often removes the wrong points. The reason is that the point ranges
are removed sequentially, and each removal changes the curve topology,
invalidating subsequent ranges. The solution is to combine the ranges
into a single index mask for each drawing before removing points, so
that each drawing is only modified in a single operation.

Pull Request: https://projects.blender.org/blender/blender/pulls/142834