Initial build support for the iOS/iPad platform
The existing known issues are presented on #142346, and include:

* No convenient `make ios` command.
  * Building instructions are tricky, the only way to build now is from
    Xcode and knowning what to configure manually.

* Missing viewport/UI refresh
  * This issue was introduced as part of the buffer swap workaround to
  fix some UI glitches :/).

* No multi-finger tap gestures (will be committed separately).
* OpenSubdiv not working (required for Release build).
* No file system support (demo files need to be bundled with the .app).
* Access to the extensions platform not working (Python multi-processing
  doesn't work on the platform).

This work is part of the Beyond Mouse and Keyboard project, and aims at
support running Blender on new computing devices. For tablets a lot of
the work is platform agnostic, though this commit focus on iOS specific
changes.

This is intended to be kept on a branch until more design and
development resources can be dedicated to it.

The commit migrates existing .icns files to Git LFS, and the sanity checking
script rejects it for some reason (it sees the deleted binary file perhaps?(

override restrictions

Ref #142346