Rename `remote_library_index` to `remote_library_listing`
Because "index" is an overloaded term, we started to name the list of
assets available on a remote asset library "remote asset listing". So now
the code also uses that term.

Filenames and functions still use "index", that'll change soon too. For now
it's just the directory name.

No functional changes.