Sculpt: Emit size for geometry sculpt undo steps
The Sculpt Undo system does not provide an estimated amount of allocated
memory when the `Geometry` type is used. This is commonly used for
non-brush operations that may affect the topology of the base mesh. As a
result, the Memory Limit undo preference is effectively ignored,
potentially leading to scenarios where much more memory is used than the
user desires and / or crashes due to OOM issues.

To address this, use the `MemoryCounter` abstraction to provide an
approximation to the overall undo system.

Note that this does not change any default behavior, as we only limit the
number of steps, not the total memory used by undo.

Resolves: #143257

Pull Request: https://projects.blender.org/blender/blender/pulls/143756