Fix #142046: EEVEE: Deferred Lighting shader doesn't run (Nvidia)
The issue in #142046 is caused by the deferred lighting pass not
running.
Despite plenty of testing and debugging we have not been able to spot
any errors in the Blender code.
Checking the link status of specialization shader variants seems to fix
the issue, which may suggest a bug on the driver side.

Pull Request: https://projects.blender.org/blender/blender/pulls/144150