Geometry Nodes: Change closure color, reset repeat zone color
Follow up to f646948ace6232fc1b8e6e164697bf26109550a7
The repeat zone color is changed back to the color it used in 4.5, to
avoid the need for people to re-learn the color. The closure socket and
zone share the same color, which is made more yellow-ish to distinguish
it from the bundle socket color.