Update online status after asset download finished
Let the downloader send a "ping" to notify about newly downloaded
assets, which the asset browser can use to transform online assets into
regular on-disk assets (if now found on disk).

To do this the asset browser iterates all assets of the library and
checks if the online asset is now available on disk. Such disk queries
are annoying to do on the main thread, but should be quite cheap.

Includes blender/blender!144200.