iOS Simulator: Skip Metal feature assert to provide basic support
This patch fixes a startup crash due to a Metal feature check assert on
the iOS Simulator. This assert is expected due to the smaller set of
supported GPU features of the Simulator. To provide basic support for
this extra platform, this assert is skipped, allowing for UI development
and testing without GPU-dependent features

As explained in Metal iOS Simulator docs[^1] fully supporting the iOS
Simulator would require large changes to the Metal backend to include
feature checks and fallbacks, as the iOS Simulator does *not* emulate
the GPU, and provide device capabilities similar to the ones of an Apple
family 2 GPU.

Due to this, when running and debugging Blender in the iOS Simulator,
certain GPU-dependent features such as Cycles GPU and EEVEE are expected
to not work, and Metal API Validation needs to be disabled.

[^1]: https://developer.apple.com/documentation/metal/developing-metal-apps-that-run-in-simulator?language=objc#Treat-Simulator-as-a-special-device

Pull Request: https://projects.blender.org/blender/blender/pulls/144095