Fix #144054: EEVEE: Wrong Attribute loading caused by multiview
This was only affecting the shadow pass because it is the
only one to use multiview.

The `drw_ResourceID_iface.resource_index` was incorrectly
set up in the vertex shader to the already shifted index,
resulting in double shift which resulted in resource index
being 0 in fragment shader.

Candidate for backport to 4.5 LTS

Pull Request: https://projects.blender.org/blender/blender/pulls/144341