Fix #144392: LineArt: End job explicitly when not baking in background
Previously the baking job is not properly ended when not baking in
background, this means when the bake is called from the script, the
interface lock is not properly reset, then parts of the interface can
stay locked afterwards. This fix adds `lineart_bake_endjob` to the
foreground code path so the interface will be unlocked properly.

Pull Request: https://projects.blender.org/blender/blender/pulls/144405