Fix #144542: Crash Workbench DoF
Attaching mipviews directly on a framebuffer results in a driver crash.
Letting the backend handle the mipviews works around the issue.

This fixes the crash, but seems to point to an issue when attaching
mipviews directly. It also ensured that the shader bound mipviews
cannot be destroyed when new mipviews are ensured.

Pull Request: https://projects.blender.org/blender/blender/pulls/144796