Render: Bake only to explicitly selected images
When baking to textures, bake only to  selected (and active) images, as
opposed to all active images.  previously the targets were all active
images in all materials. This made it very unclear which images were
baked to, as images can be active without being selected. It also made
it impossible to *not* bake to any image in a material since there is
always an active image as long as an image texture node exists. This
often lead to accidentally overwriting images of existing PBR materials.

Baking only to selected images fixes the workflow where you have a model
with multiple materials where you only want to bake some of them. For
example a model with a PBR material for the roof and a procedural bricks
material for the walls. On export you want to bake the procedural
bricks, but not the  PBR roof.

This is an API breaking change.

Pull Request: https://projects.blender.org/blender/blender/pulls/137389