Fix #144636: EEVEE Invalid hit during raytracing
Raytracing pipeline in EEVEE interpolates in time for better results.
But depending on the interpolation could try to load a sample that isn't
available.

This PR fixes this for planar probes and screen tracing. In screen
tracing the issue wasn't visible as it the background samples are
always cleared.

This has an expected performance penalty but it could require an
overhaul of EEVEE raytracing for a more efficient solution.

Pull Request: https://projects.blender.org/blender/blender/pulls/145508

Pull Request: https://projects.blender.org/blender/blender/pulls/145545