UI: Dyntopo Confirmation Improvements
Enabling Dynamic Topology can result in a warning message when there
are existing attributes or modifiers. Unfortunately this warning looks
like an alert (single OK button) but is actually a confirmation. That
it operates when you click the single button is not clear. Nor is it
clear how to cancel. This PR replaces this with the current style of
confirmation, with clear execution and cancel buttons.

Pull Request: https://projects.blender.org/blender/blender/pulls/145534