Ignore old pages while loading asset listing
When requesting a listing (re-)download, store a timestamp (using a
C++17 file system clock, so the precision matches file time stamp
precision), and use that to ignore previously downloaded pages. The
downloader already touches listing files as soon as they are known to be
up-to-date.

The downloader currently doesn't touch the `asset-index.processed.json`
file when up-to-date, this should be fixed still. For now commented the
check for its timestamp out with a TODO comment.

Also ensure data of the download status class is initialized.