Disable asset dropping while still downloading
Dragging an asset triggers downloading. Now when dragging it shows the
asset preview and name next to the cursor as usual, but with a
"Downloading asset..." text. Dropping is disabled (the cursor indicates
this) and won't do anything, until the asset is done downloading. Then
the "Downloading asset..." disappears.

This prevents a crash from trying to import an asset that's still
downloading.