Win: Improve crash reporting
When reporting a crash with unknown exception we just display
UNKNOWN EXCEPTION which is not as helpful as it could be. this
commit adds the actual exception code for all reports.

Also adds a msvc specific exception to the known list for easy
identification.

the exception record from #145762 goes from

ExceptionCode         : UNKNOWN EXCEPTION

to

ExceptionCode         : Microsoft C++ Exception (0xe06d7363)