Fix #145753: Crash deleting a keyframe for another slot when tweaking
When deleting the last key of an F-Curve, the F-Curve itself also gets
deleted. This is now done via the same code as deleting F-Curves from
the channel list.

Before this fix, a deletion function was used that's not capable of
deleting from Actions in NLA strips, which is why it crashed.

Pull Request: https://projects.blender.org/blender/blender/pulls/145929