Fix #145715: Right click select doesn't work with Blade Tool
Blade cursor patch #136749 cancelled select operations if the current
tool was the blade tool. This was to prevent LCS users from being able
to select handles when the blade tool would pass-through.  This however
resulted in an RCS regression.

Fix instead by cancelling split if the mode is `use_cursor_position`
and there is no valid split, preventing a pass-through.

In RCS this works nicely because users can just select using right-click
and bypass split logic. In LCS this makes it impossible to select with
the blade tool enabled, but this is expected behavior as the blade tool
does not need strips to be selected to work anyways, and tools should
ideally have different, non-overlapping purposes. For instance, the slip
tool also does not affect selection logic.
(Note: there is one small caveat to this principle, which is that the blade
tool can't currently split connected strips together unless theyre both
already selected. This issue should be fixed with a future blade tool
refactor).

This also fixes unrelated bugs:
- Blade cursor no longer shows on first/last frames, where split point
  fails
- You can no longer split in the gutters between channels

Pull Request: https://projects.blender.org/blender/blender/pulls/145938