Fix: Grease Pencil: Assert when converting to `Catmull Rom` or `Poly`
These `BLI_assert`s could be triggered by converting curves that were
already the desired type.

`curves.is_single_type` Checks if all of the curves are the desired
type, not if all of the selected curves are the type.
So the asserts can be triggered when only the destination type curves
are selected and other curve types exist.

Pull Request: https://projects.blender.org/blender/blender/pulls/145529