Fix #141142: UI: Emboss status not restored in uiLayout
When using `emboss=False` on a UI prop, if `property_split` was
enabled, all subsequent props in the same layout would also be
unembossed.

This is because of the way the emboss status was set.

When drawing a prop, the previous emboss status was stored in a
temporary variable `prev_emboss`. This status was retrieved from the
uiLayout (`this`), but restored to a `layout` uiLayout that could be
another one, because it could be reassigned in the mean while.

Instead the status should just be restored to `this`, otherwise the
layout can keep the same unembossed status for subsequent props.

Pull Request: https://projects.blender.org/blender/blender/pulls/145883