Win: Preliminary VS 2026 support
There is no CMake version yet that support VS2026, support is
currently limited to ninja based builds.

CMake does not know where to look for the MSVC 2026 Runtime
it copies the wrong CRT leading to startup issues, for this reason
WITH_WINDOWS_BUNDLE_CRT has been disabled for VS2026 + CMake < 4.2.0.
Builds done with VS2026 are for this reason *NOT* distributable to
end users at this point in time.

CMake 4.2.0 Is *expected* to have VS 2026 support [1]

But further testing is likely required.

Things of note:

- They stopped using the "preview" channel and have rebranded to
"Insiders" the make.bat options have been updated accordingly to
select the 2026 Insiders build, you need to pass 2026i to make.bat

- All tests seem to pass on my system

- No new warnings, pretty smooth update

[1] https://gitlab.kitware.com/cmake/cmake/-/merge_requests/11168