Fix: Curve and Line stroke types can have incorrect spacing behavior
The intent of the Curve and Line stroke types is generally to perform
a set of equally spaced points. Unlike drawing with a tablet using the
Spacing stroke type, pressure should have no influence on this input.

Previously, the code would pass in a hardcoded value of 1.0f for the
line & curve pressure. This is incorrect, as it results in a factor of
0.5f, not 1.0f, causing behavior to differ on this stroke type depending
on if the corresponding brush pressure flag was enabled or not.

Additionally, an extra check is used to ensure that only the Spacing
stroke type can have pressure modulated in this way.

Pull Request: https://projects.blender.org/blender/blender/pulls/146120