Fix: Artifacting in experimental texture paint on non-manifold meshes
When texture painting a non-manifold mesh, various artifacts are created
on the image texture.

To add seams to the edges of non-manifold UV islands, the texture
painting system reads and writes pixels near the edge of the UV islands.

The offset calculations used to find pixel positions for reading and
writing were not accounting for color channels (RGBA) in the pixel data.
Because of this, the reading function corrupted the color data for the
seams, and the writing function caused the seams to be scaled down and
repeated.

Pull Request: https://projects.blender.org/blender/blender/pulls/146110