OpenEXR: Support reading multipart files with full channel names
The reading code currently assumes that the part name may be for example
"ViewLayer.Combined" and then the channel within the part may be "R". For
example Houdini writes files like this.

However according to the docs the part name should be ignored and the
channel name in each part should be complete like "ViewLayer.Combined.R".
https://openexr.com/en/latest/MultiViewOpenEXR.html

To support both cases we now only prepend the part name if it's not already
there.

Pull Request: https://projects.blender.org/blender/blender/pulls/146650