Avoid redundant directory recursion in Find Missing Files operator
With the previous implementation, the code would recurse through the given
directory tree for each missing file.
That's not necessarily a cheap operation (e.g. on shared network drives where
metadata may not be cachable), but since we look at the same directories for
each file, so we can avoid the repetition by a) building a map of files that
we're looking for, b) recursing through the directory and checking each entry
for whether it's in the Wanted list, and then c) going over all missing files
and replacing their path if we found it.