Compositor: UI: Inline sockets
Where possible, inline sockets when the input and output are both the
image or the same value just transformed.

Even though it may make sense, do not apply this change in cases where
the image is the same when a factor is set to "0". For example the
Alpha Over node, where the "Background" matches the "Image" output if
factor is 0. Or the Mix node's "A" and "Image".

Also hide the value (usually a full white color input widget).

It makes for a more compact UI, less noise, and it matches a similar
layout in Geometry Nodes.

See PR for screenshots.

Pull Request: https://projects.blender.org/blender/blender/pulls/147143