Fix: Adjust adaptive subdivision versioning to set defaults in all cases
When an old file that uses regular subdivision is loaded, the adapative
settings were not applied to the existing modifiers if they weren't
already using adaptive subdivision. This results in the settings
remaining at their 0 defaults which can be surprising. E.g. the
`dicing_rate` of 0 is clamped to 0.1px which is probably not intended.

Make sure to set the correct defaults in all cases now.

Pull Request: https://projects.blender.org/blender/blender/pulls/147164