Fix #147260: Freestyle mark not saved in files in some cases
When there was no other data written in the CustomData for edges
or faces, the layers pointer would be null, and even though the layer
array we write isn't empty, the writing would be skipped. The fix
is just to make sure the layer pointer isn't null.

Pull Request: https://projects.blender.org/blender/blender/pulls/147411