Nodes: simplify editing group interface based on selected group input/output node
This fixes two annoyances when working on node groups: * One has to constantly
switch between the Group and Node tab in the sidebar. * Finding a specific group
input/output in the tree view can be annoying when there are lots of inputs.

This patch implements two features which simplify the process significantly: *
When selecting a group input/output node, activate the corresponding item in the
tree view.
  * When the node has multiple sockets, use the one that's closest to the
  cursor.
* Show the group interface tree view in the node properties when selecting a
Group Input/Output node.

This constains one UI change that may affect other areas: When a new tree view
is added on a redraw, the active element is now scrolled into view, instead of
the scroll bar just being at the very top.  It seems generally good to me, but
could be made optional.

Pull Request: https://projects.blender.org/blender/blender/pulls/147365