Fix #146242: Crash when aligning bones of different armatures
When trying to align bones of 2 different armatures, with different mirror settings
it would crash.
The reason it crashed is that a pointer was not set to anything.
The whole function is not set up to do multi object editing and assumes that all bones gotten
from `selected_editable_bones` are part of the same armature. This is explained by the
comment above that explains that any unselected bones returned from this function are
assumed to be the mirrored bones.

Pull Request: https://projects.blender.org/blender/blender/pulls/147521