Fix #147786: Crash when connecting menu to float input
Compositor crashes when connecting a menu output to a float input, this
is because no implicit conversion is supported between the two, and no
handling was done in this case. To fix this, we fallback to a default
value for unsupported implicit conversions.

Pull Request: https://projects.blender.org/blender/blender/pulls/147801