Fix #144641 (for 4.5 LTS): Rigify: Assign first slot for Action constraints
Ensure that the first suitable action slot is used for Action
constraints.

This is the minimal fix to this issue for 4.5 LTS. Surrounding code
was written before Action Slots were introduced, so it was generating
Action Constraints without selecting a slot, which were invalid. We
just assign the first slot of the chosen actions, so it behaves as it
did before Action Slots existed.

-----------

This is just a fix for the 4.5 branch. The proper fix for 5.0 is in !146182.

Co-authored-by: Demeter Dzadik <metemer001@gmail.com>
Co-authored-by: Sybren A. Stüvel <sybren@blender.org>
Pull Request: https://projects.blender.org/blender/blender/pulls/146287