Asset Downloader: keep AssetDownloader instance around after shutdown
Earlier, the AssetDownloader would be completely deallocated when its
job was done, and its last-known status was stored separately. This made
it a bit cumbersome to reason about, and also harder to duplicate the
logic for thumbnail downloading.

With this change, the AssetDownloader's BackgrounDownloader still shuts
down, but the AssetDownloader itself is kept around and is 'restartable'.