Download asset previews on demand
Instead of downloading all previews when the listing is downloaded,
download them on-demand from the asset browser (or, whenever Blender
asks for them).

For now, thumbnails are only re-downloaded after 7 days, to avoid
hammering the web server. This is purely determined by the local file
path of the thumbnail, and since the local file path is determined by
the asset and not the thumbnail URL, thus is insensitive to URL changes.

Future improvements:
- Store the thumbnail URL somewhere, and always re-download when the URL
  changes. This makes it possible to put version information in the URL
  by updating the remote listing, giving asset authors a form of "cache
  busting".
- Add support for the `Cache-Control` HTTP header. This can specify a
  max age, but also would require the asset author to understand HTTP
  well enough, and have enough control over the web server to modify
  this.