Cleanup: UI: Reduce variable scope and use const in View2D function
There's a lot of scrolling needed to see how variables in this function
are used and if/where they are modified. Declaring them later (closer to
where they are used) reduces need for scrolling. Making variables const
also makes clear that these are not modified between declaration and
use, reducing the need to search for variable uses/assignments when
trying to follow the function logic.