Fix: VSE: Crash when rendering scene strip that points to same scene as VSE timeline
When a "sequencer-type" scene strip inside of scene A pointed at another
VSE timeline in scene B that had a scene strip that pointed back to scene
A, and a final render was triggered (with other renderable strips in scene A's
timeline), Blender would attempt to create an identical render depsgraph
with the same scene and bmain as the original render depsgraph
leading to a crash.

This fixes the issue by simply adjusting the recursive check that
already existed for "sequencer-type" scene strips, so that the check
also applies to any scene strip. Note that now the marked scene is the
enclosing `context->scene` rather than `strip->scene` (so that we can
immediately notice a scene recursion).

More info in PR.

Pull Request: https://projects.blender.org/blender/blender/pulls/147955