Fix #147996: Shift selecting in the outliner can cause the Dope Sheet to only show the summary
When Shift selecting a range of items in the outliner, all other items are deselected.
This also deselects the armature object causing the anim filtering code to
skip it if the "Only Show Selected" flag is enabled.

There is already functionality in `animdata_filter_base_is_ok` to check cases
where the object is in pose mode and ignore the selection state.
That was introduced in 2c2e1b3d612d87b54ce3c3f1e52b7c89ee88a5c3
However `animdata_filter_ds_sorted_bases` always hardcoded the object
mode to `OB_MODE_OBJECT`.

The fix is to take the mode from the given `Base`.

The reason this doesn't happen with Ctrl select is that in that case the object
is set to be active and selected.
I could have tried to fix the outliner selection behavior which you could argue is
buggy, but since I found that that function `animdata_filter_base_is_ok`
already has code that is supposed to deal with this case I decided to keep it
local to the animfilter system.

Pull Request: https://projects.blender.org/blender/blender/pulls/148491