Fix #147810: Node outline missing theme setting
The subtle outline introduced in 916f0afd45 is too strong in light
themes, with users reporting eye strain (#147810).

Add a theme setting instead so it's possible to adjust it or make
it fully transparent. Similar to panels.

This is also an improvement in accessibility because contrast can be
enhanced in regular nodes, not just the active and selected ones.

See PR for screenshots.

Fixes #147810

Pull Request: https://projects.blender.org/blender/blender/pulls/148532