PyAPI Docs: Fix "On this page" covering content
Sphinx CSS "hides" the _On this page_ pane by setting its `position` to `fixed` and `right` to `-15em`—this assumes it has a fixed `width` of `15em`, which it doesn't if it is overridden to `auto`.

Fixes #148347; caused by [45fee7d851][0], which also affects the 4.5 LTS docs.

[0]: https://projects.blender.org/blender/blender/commit/45fee7d8517dfbb508d578fdf8c45f85865ade9c

Pull Request: https://projects.blender.org/blender/blender/pulls/148603