Fix #148562: Overflow when uploading pointclouds to GPU
An overflow can happen when a huge amount of points are drawn. Each
point is triangulated and the index buffer can only store 4GB at max.
GPU index buffer API uses int/uints for size calculation.

During upload it is now detected if the index buffer could safely be
stored on the GPU. If this isn't the case it will draw no points for the
pointcloud.

Other improvements can be made in this area to provide better feedback to
the user. This only fixes the write out of bound and crash.

Possible suggestions are:

- Display a message in the GN/node/editor.
- Separate into multiple vertex/index buffers to relax the limitation
- Replace index buffer with a vertex shader

Pull Request: https://projects.blender.org/blender/blender/pulls/148631