Fix #146255: Scrub region numbers don't appear with lines
The issue was that the lines and the labels for those
did not appear at the same time. The reason for that
was that there is a padding that has to exist between
labels. The lines appeared at a point where the labels
would have violated that padding.

The fix is to shrink the padding slightly.

Also fixed in this patch is the function to get the string
from the numbers not respecting the UI scale.

Pull Request: https://projects.blender.org/blender/blender/pulls/148649