Stop background downloader process when Blender dies
When Blender dies in a non-clean way (which prevents it from shutting
down background processes gracefully), this is now detected by the
background downloader process, and makes it shut down.

This is done by catching the `EOFError` and `BrokenPipeError` that will
be thrown when receiving resp. sending messages from/to the main
process.