Different way of notifying asset browsers about downloaded previews
The previous way of using the message bus wasn't working entierly yet (a
single downloaded preview would cause all visible previews to be pushed
to the "to load from disk" job queue again and again) and needed more
work. However, I don't see a need for such a solution, this is much
simpler, works reliably, and is easy to follow in code.

Now only previews that are finished downloading get put on the "to load
from disk" queue.