Linux: Preload system TBB if it is compatible and newer
This avoids issues where the Intel GPU ray tracing driver may depend
on the newer system TBB version. In particular, on Ubuntu the TBB
library is no longer statically linked into the driver as it was before.

We think this really should be fixed upstream in distributions, and
that graphics drivers should not pull in TBB as a dynamic library. As
other VFX software will run into the same issue, and this breaks an
implicit convention that graphics drivers avoid such ABI conflicts.
But policies around static and dynamic linking make this difficult.

This fix only affects the Blender Snap package and the blender-launcher
script used for Steam, to reduce the risk for Blender 5.0. For later
versions we may change the Blender executable itself into a shell script,
or use an alternative fix that makes Blender into a shared library.

Ref #140631

Pull Request: https://projects.blender.org/blender/blender/pulls/148154