Fix: Improper duplicate material handling in USD import
Part 1/2 for #148569

The map used to de-duplicate incoming materials was filled with the
wrong material names which eventually resulted in a situation where we
were adding 2 different materials with the same name to the map which is
unexpected.

This also contains a fix for Material ID user counts that was
incorrectly changed as part of the unrelated refactor 1b4daf9d2e081562d.

Added further test coverage to validate both situations now.

Pull Request: https://projects.blender.org/blender/blender/pulls/148757