Fix #145853: Crash using Blade or Slip tool in VSE with new scene (4.5 backport)
If adding a scene from the main window header, it is possible to have a
defined `scene` but an undefined `ed`.

The `tref->idname` checks may be removed since there's no
other kind of tool other than select / blade so the check always
evaluates to false anyways.

Technically in this version of blender (4.5) it is usually impossible
to have a `nullptr` scene but I have included the check here for
consistency with the original 5.0 fix.

Pull Request: https://projects.blender.org/blender/blender/pulls/148175