Fix #141186: File open dialog resolves symlinks on macOS
It was detecting aliases as symlinks, but we only need to resolve actual
aliases as they are not recognized by common file system APIs.

BLI_file_attributes on macOS now sets the symlink flag, and disables
the alias flag if it's a symlink.

Pull Request: https://projects.blender.org/blender/blender/pulls/148718