GPU: Add names to GPUFences
GPUFences are not named. During debugging a developer
needs to walk along the stack frames to find out which
fence is being used. This PR adds a name to the GPUFence
object.

For Metal the event object will be named. In future the Vulkan
fences might become VkEvents and can use this name as well.
Ref: !148720

Pull Request: https://projects.blender.org/blender/blender/pulls/148721