Fix #148762: image inputs not working in node tools
Data-block sockets are stored as strings in node tools currently and that's not
supported by `uiTemplateID` currently. This uses the same `prop_search` that is
used for other data-block types for images for node tools. The modifier still
uses the template which shows the new/open buttons.

Separately, `uiTemplateID` could be improved but that's a bit more involved.

Pull Request: https://projects.blender.org/blender/blender/pulls/148884