Refactor: Use attribute API for data transfer UVs and colors
Accomplish this by storing either void pointers or C++ containers
in the layer map struct with a variant. This is a more proper thing
to do because it tags attributes changed, though that's currently
unused for such generic color attributes. But the main purpose is
to avoid changing this code when Mesh transitions away from using
CustomData to store generic attributes.

More of the code accesses attributes with names now, where previously
it used the index in the CustomDataLayers of a specific type.

Part of #122398