Fix #142864: Cannot grab playhead in invisible marker region
This issue happened as a combination of 2 fixes
* cfc42f7f546fe1cab9c242b335ab957114702b00 which introduced the fading
* 4306e746a5d which masked out events in the marker region

Since the marker region was no longer visible,
it got very confusing as to why the hotkeys no longer work.
My fix is to remove the fade of the marker region,
especially because at the default height of the timeline
in the startup file the marker region is completely invisible.

I chose not to pass through the events.
This patch just makes the marker region background *not fade* so it is clear
that it exists and if hidden ensures that events are not blocked.

Pull Request: https://projects.blender.org/blender/blender/pulls/148633