Vulkan: Compute Pipeline Compilation Improvements
This improves the compute pipeline compilation. The previous implementation
did a lock when compiling the compute pipelines which removed the ability
to compile multiple pipelines at the same time.

This change only locks when the pipeline cache is being read or altered.
It also has the ability to wait for another thread that is compiling the exact
same pipeline.

After this lands in main the graphics pipelines will also be refactored, but
as it has a much higher risk we will do that in a separate PR [!148843].

Pull Request: https://projects.blender.org/blender/blender/pulls/148499