Fix #147719: `Shift + Tab` hotkey missing for Dope Sheet and Graph Editor
This patch adds the missing `Shift + Tab` hotkey for toggling
transform snapping to the Dope Sheet and Graph Editor.
This makes it consistent with other areas of Blender.

This was discussed in an Animation & Rigging module [1]
and it was decided to implement this.

[1] https://devtalk.blender.org/t/2025-10-16-animation-rigging-module-meeting/42949

Pull Request: https://projects.blender.org/blender/blender/pulls/148987