Refactor: Vulkan: Create Graphics Pipeline
The construction of the graphics pipeline is a large function. To make
the function easier to read it was split in multiple smaller functions.
This isn't a cleanup as it touches quite a large amount of lines and
restructures how data is managed.

Pull Request: https://projects.blender.org/blender/blender/pulls/149051