Fix #149062: VSE: Correct meta strip end point detection
In `sequencer_meta_make_exec`, it scans the end of the meta strip with
the minimum frame number of 0, which can't handle the case where the
strip is completely put in negative range of the timeline. Now the end
point starts from `INT_MIN`.

Pull Request: https://projects.blender.org/blender/blender/pulls/149065