Fix #148457: Movie fallback seeking not working correctly
For MTS file format fallback seeking is used and with specific files
`av_seek_frame()` function fails to seek to specific timestamps. As a
result, packet, that is read matches previous one. This was previously
interpreted as attempt to seek to time before stream start, as the
behavior would be the same. Use stream start timestamp explicitly to
allow for slightly broken ffmpeg behavior.

This issue seems to be fixed in ffmpeg master branch

Pull Request: https://projects.blender.org/blender/blender/pulls/149102