Geometry Nodes: Support for baking lists of primitive values and bundles
Adds support for baking List items in simulation zones and bake nodes.
In addition to primitive value lists this also supports lists of bundles, which
are not currently accessible from the user level. Lists of bundles are still
used in the physics solver branch for debugging and should be fully supported in
the UI in the future.

Pull Request: https://projects.blender.org/blender/blender/pulls/148958