Fix: UI: Tree-view search section expanded by default in existing files
To have the section collapsed by default in existing files with the "collapse
filter options" flag, we'd have to set the flag in versioning. However, we can
just invert the flag to mean "show filter options" and avoid having to version,
since it's not set by default then. Also makes semantics easier to follow and
avoids negations in code.