Nodes: Node Wrangler: Add "Center Nodes" operator
__Problem__
No quick and easy UI element to help organize nodes at or around the center of the node editor.
[Extension platform submission with similar feature](https://extensions.blender.org/approval-queue/node-position-reset/)
There are also a couple of RCS requests to add axis lines to the node editor for visual aid.
https://blender.community/c/rightclickselect/E6Rm
https://blender.community/c/rightclickselect/3wNp

__Solution__
This PR adds an operator for moving selected nodes to the center of the node editor.
This does not directly solve the problems mentioned in RCS threads requesting axis
lines, but still provides a quick way to move nodes to, and find the center of the node editor.

The operator will run on the outermost selected nodes
- If selected nodes are not inside a frame, move those nodes.
- If selected nodes are inside a frame and the frame is also selected, move the frame.
- If selected nodes are inside a frame and the frame is NOT selected, only move those nodes.

Image and video in the PR.

Pull Request: https://projects.blender.org/blender/blender/pulls/148286