Fix: Metal: Invalid uniform setup of image and sampler in shader interface
These resources do not have mutable binding points.
This is unfortunately the case in GLSL. Even if we
don't allow such practice the pyGPU still sets the
sampler bindings using the uniform API to ensure
valid values in OpenGL.

This led to this call overriding the 1st member of
the push constant buffer. If this component was
part of a matrix, it could lead to broken rendering.
This was the case for the builtin `IMAGE` shader.

Setting their location to -1 discard the uniform
function call and fixes the issue while keeping the
opengl behavior unchanged.

Thanks @brita for making me spot the bug with your
addon.