Fix #146795: Grease Pencil reproject needs to use per-thread snap context
The `transform::SnapObjectContext` is used to store intermediate results of
raycasting, such as the closest point. It needs to be thread-local storage
to avoid mixing up results of parallel threads.

Would be nicer if the "context" part could actually be a separate immutable
context, but this is old code and not worth fixing now.

Pull Request: https://projects.blender.org/blender/blender/pulls/149258