Fix #143853: NLA apply scale was no longer working
The nla.apply scale operator was using the old way of iterating
the action, i.e. not using slots. This is now fixed and by doing so
the whole function was no longer used so I removed it.

Note: needs to be backported to 4.5

Pull Request: https://projects.blender.org/blender/blender/pulls/148989