Nodes: Don't darken color picker of inactive sockets
When Color sockets are not used, they appear grayed out.
Although this is helpful to show that the input is irrelevant, it might
be confusing to show a darker input color than what the socket really
contains.

To fix this, only the label gets grayed out if the socket is inactive.
See PR for

Pull Request: https://projects.blender.org/blender/blender/pulls/149083