Fix: Vulkan Validation Error
When WITH_DRAW_DEBUG is on the debug shader is compiled. This is the
only shader that writes from the vertex shader and requires the
`vertexPipelineStoresAndAtomics` feature to be turned on.

It is safe to turn it on as all our platforms supports this feature.
There is no 'real' need to backport this as the `WITH_DRAW_DEBUG` is
a developer option.

Pull Request: https://projects.blender.org/blender/blender/pulls/149293