Fix: Movie Distortion node introduces an offset
The Movie Distortion node introduces a one pixel offset even if the
distortion model is identity. This is because shrinking bounds were
computed using flooring, while it should be ceiling, because shrinking
is a negative delta, which we need to reduce to avoid data loss at the
edges.

This also exposes an issue where data loss happens at the edges since we
only consider the coverage of pixel centers, while we should consider
the pixel edges as well.

This fixes part of #146246.

Pull Request: https://projects.blender.org/blender/blender/pulls/149336