Fix #148073: Crash entering sculpt after deleting base multires vertices
Modifier code did not handle a mesh with loose vertices & edges
correctly. As part of creating the `SubdivCCG`, we require the
`TopologyRefiner`, which can be null despite having a valid `Subdiv`
in the case that the underlying Mesh has no faces.

To avoid crashing, check for this case and return an empty but
constructed `SubdivCCG`. Note that this still does not allow sculpting
on the loose elements, but that case will be handled separately and
likely requires more work.

Pull Request: https://projects.blender.org/blender/blender/pulls/148093