GPU: Add GPU_shader_batch_pause/resume_compilations
Follow the implementation requested in 148933[^1].

Implement functions to pause and resume shader compilation globally, so
when cancelling a large number of compilations (e.g. during app exit at
the moment, but possibly on scene change or render cancel in the future)
we don't run into the issue where cancelling one shader that is already
being compiled makes the calling thread wait after the next shaders have
already been scheduled for compilation, causing a chain of waits and
making the cancellation slow and ineffective in practice. (This issue
was introduced in #148933 to prevent use-after-free crashes.)

The main downsides of this method are that it requires extra handling
from the calling code for the cancellation to work as intended, and that
it exposes control of the internal scheduling to the public API, which
may cause issues when multiple threads use this or other
`GPU_shader_batch` API calls at the same time. Some asserts have been
added to detect incorrect uses of the API.

Other options, if this becomes an issue in the future, could be:

* Ref-counting the `ShaderCreateInfo` so no wait is needed.
* Passing the batches we want to cancel at once, so scheduling can be
managed internally.
* Adding a `try_cancel` function, so we can run `try_cancel` for all
shaders before running the blocking cancel.

Targeting `main`, since dealing with multithreaded synchronization
always has the potential of introducing race conditions and deadlocks.

Note there is a pending refactor to use the `BLI_thread_queue` API for
`GPUWorker` (#140993), which may be relevant for the code in this PR.

Pull Request: https://projects.blender.org/blender/blender/pulls/148957

[^1]: https://projects.blender.org/blender/blender/pulls/148933#issuecomment-1746650