Update the AUTHORS file with the orgs from credits_git_gen.py
For the orgs that I did not find a contributor with a valid e-mail I
used <> (which is still a valid pattern, it just won't match anything).

I'm also adding PICO to the list (the contribution was from hogan and the
email is from the parent company - ByteDance).

Pull Request: https://projects.blender.org/blender/blender/pulls/149423