Subdiv: Use attribute API for CPU attribute interpolation
As part of #122398, change the subdivision evaluation code that produces
a mesh to use the attribute API rather than using CustomData directly.
Mainly this is to facilitate not using CustomData to store generic
attributes, but the code is improved in a few miscellaneous ways in
the process:
 - There is no more allocation per non-quad face anymore. That was done
   to build a CustomData struct with 4 elements for every new face. Now
   the same per-thread arrays are used for the whole process.
 - The complexity to avoid interpolating face corner topology attributes
   isn't necessary anymore. We can just avoid interpolating specific
   attributes by not including them in some local vectors.
 - We no longer do redundant interpolation of positions and UV maps.

The code is still structured in a quite inefficient way, in my opinion,
mainly because it's built around multiple different callbacks per
topology element. That results in a lot of branching that will mean
the hot loops are not very efficient.

This change should make it easier to do further data-oriented
refactors to this area and that can be continued in the future.
For now this resolves the immediate need for replacing CustomData.

Pull Request: https://projects.blender.org/blender/blender/pulls/149152