Anim: Increase performance of BKE_fcurve_bezt_binarysearch_index_ex
This is a small change that improves the performance of
`BKE_fcurve_bezt_binarysearch_index_ex` by a small amount.

The binary search had an additional unneeded `if`.
It wasn't necessary because equality is checked before, then `>`
so the only remaining option is `<`.

Because the function is so low level this has a measurable impact on performance.

On the synthetic test the frame time improves by 3.4ms which translates
into a FPS change from 33 to 38 fps.
The rain file increased by ~2fps.
In general: The more FCurves there are the higher the gain.

part of #149405

Pull Request: https://projects.blender.org/blender/blender/pulls/149450