Refactor: Compositor: Check valid compositing region engine side
This patch moves the checks for the validity of the compositing region
from the various operations to the engine side, where the engine simply
does not run in those cases. This only needs to happen for the viewport
compositor because other front-ends already handle this.

Pull Request: https://projects.blender.org/blender/blender/pulls/149632