GPU: Shader: Move shared variable defnition to create info
This commit parses the shared variables in global scope,
erase them, and move them to the create info.

The backend is then in charged of defining it properly
in the target language.

This creates some new limitation:
- Type must be trivial (eg. float, int, ...).
- Type must not be a macro.
- Array size must not be a macro.

This is needed to simplify codegen in the metal backend #147010.

Pull Request: https://projects.blender.org/blender/blender/pulls/147084