Fix #149671: Tag geometry caches dirty when attribute is changed via Python
The current attribute API is somewhat bad currently for various reasons and a
better alternative is occasionally worked on already.

One of the issues right now is that certain geometry caches are not cleared when
attributes are changed. Ideally, we could detect exactly which attribute is
modified, but this is not really doable with the current implementation. This
patch fixes it by simply tagging all caches as dirty when any attribute is
changed via RNA. This often tags more than necessary, but should still be
correct. A new attribute API would handle this more cleanly.

This makes it even more beneficial to use `foreach_set` when changing multiple
attribute values.

Pull Request: https://projects.blender.org/blender/blender/pulls/149710