Compositor: Make context IO operations display window aware
This patch makes the context input and output operations aware of the
display window. This currently only concerns the viewport compositor
when in camera view, otherwise, the data and display windows are the
same.

This has no functional changes yet, as individual operations would need
to be adjusted to take the offsets and display into account.

Pull Request: https://projects.blender.org/blender/blender/pulls/149762