Compositor: Make Box and Ellipse masks Display Window aware
This patch updates the Box and Ellipse mask nodes to be aware of the
display window of the input. In practice, this means that masks will
retain their position and scale even when the camera region is moved
and zoomed in the viewport compositor.

Pull Request: https://projects.blender.org/blender/blender/pulls/149773