Compositor: Make Scale node aware of Display Window
This patch updates the Scale node to be aware of the display window of
the input. In practice, this means that scale node in Render Size mode
will retain and scale even when the camera region is moved and zoomed in
the viewport compositor.

Pull Request: https://projects.blender.org/blender/blender/pulls/149840