Compositor: Make Relative To Pixel node Display Window aware
This patch updates the Relative To Pixel node to be aware of the display
window of the input. In practice, this means that properties controlled
by the node would maintain their relative size even when the camera
region is moved and zoomed in the viewport compositor.

Pull Request: https://projects.blender.org/blender/blender/pulls/149888