Images: Support reading EXR Display Window information
This patch adds support for reading the display window information that
exist in EXR files. This is needed by the compositor, which now supports
display window. Compositor-side code will be added in a separate patch
to make use of this information.

Those information are stored in the ImBuf struct and are only optionally
initialized for images that support display windows, so a flag was added
to denote the existence of such information.

They are not always initialized because the x and y members which denote
the data size are initialized in many places in the code base, so making
sure the display information are initialized correctly in all cases will
be very difficult.

A problem might arise in code that copies the image buffer flags but
doesn't copy the display information as well, in that case, the display
information could be read as uninitialized data. In practice, this will
not be a problem since those will be read as is only in the compositor
for now.

The display_offset could be utilized to represent compositor backdrop
translation instead of using special members in the image ID runtime.
This will be looked into in a future patch.

Fixes #109894.

Pull Request: https://projects.blender.org/blender/blender/pulls/149827