Refactor: Mesh: Add LegacyMeshInterpolator
To simplify the transition #122398.

This can be extended to copy and mix attributes stored in
AttributeStorage as well as CustomData, without changing all the places
it's used. The point of adding this now is to make the diff for the
final switch to AttributeStorage smaller and easier to review.

Pull Request: https://projects.blender.org/blender/blender/pulls/149876