Refactor: Mesh: Simplify attribute iterators
Instead of using a custom RNA iterator with a skip function,
build a vector of pointers when starting the iteration, and iterate
over that as a regular array iteration. Though this requires an
allocation, I'd expect that to be cheap compared to the current
quadratic-style iteration used to check when to switch to the
next CustomData struct. The main purpose of this commit though
is to reduce the amount of code that has to be ported to
AttributeStorage from CustomData.

Part of #122398