Fix #149965: Cycles missing tangent attributes
Most of the time these would work anyway because typically they are used in
combination with the UV map. But if they are used on their own, we need to
explicitly request the UV map so we can compute the attribute from it.

Pull Request: https://projects.blender.org/blender/blender/pulls/150033