EEVEE: Precompile World Material Pipelines
This PR will compile EEVEE WorldPipeline and Background pipeline shaders
when compiling the world material. This is a first step to ensure most material
pipelines can be compiled in a background thread. Currently they are compiled
when first used in the main rendering thread.

This extends ShaderCreateInfo to add descriptions of pipelines that will be
compiled. The API of this has been validated to support both Metal and Vulkan
requirements. However this PR only adds the changes to Vulkan.

Metal backend will be added in a separate commit as it has some code
duplication as it state is partly stored inside the MTLContext.

In debug builds a message will be logged when a pipeline compilation is triggered
after pipeline states have been precompiled. This can be the case when the needed
state has changed due to changes in EEVEE or when the same material is used by
object that where the GPU buffers are structured differently (missing attributes, or
multiple materials).

`GPU_context_debug_pipeline_creation` is introduced that can be used during
development to assert when a pipeline compilation is triggered. Unit tests can
temporary enable this for validation of the mechanism.

The vertex inputs API is untested and unused. It most likely need some changes
when we start using it. The changes would include a better balance between the
responsibility between the engine (EEVEE) and the backend.

Pull Request: https://projects.blender.org/blender/blender/pulls/149454