Fix #150050: EEVEE: NaN texture coordinate from Environment map
Using 0,0,0 for input texture coordinate in the
environment texture node would result in
NaN Sampling coordinate which resulted in black
pixels. This was caused by `atan` function being
fed with x = 0 which is undefined behavior with
fast math on.

This did also affect other platforms and not
only Metal.

I would consider this fix to be risky for 5.0.
Given the bug has a workaround, I wouldn't
say it is critical. So targeting 5.1.

Pull Request: https://projects.blender.org/blender/blender/pulls/150082