Sculpt: Update tooltips to propertly capitalize 'face set'
The concept of face sets are similar to vertex groups or other
attributes in that they are not brand / product names like Cycles and
Grease Pencil.

This commit changes tooltip descriptions for this phrase, as well as the
`sculpt.face_sets_init` descriptions in genral to conform to the HIG
standards.

Pull Request: https://projects.blender.org/blender/blender/pulls/150097