Vulkan: Enable building on MacOS
Allow building of the Vulkan backend on MacOS. This doesn't mean
that running will work or that we are aiming it to be a target for a
future release. Its primary reason is as a convenience for developers
to tweak the Vulkan backend and ensuring it compiles when developing
on Mac.

Running using Vulkan is a different story. It requires a build of KosmicKrisp
(part of mesa). There are features that are currently not supported. EEVEE
and OSD among others will not work due to these missing features.

Pull Request: https://projects.blender.org/blender/blender/pulls/150106