Fix #149845: Some linked objects may not be instantiated.
With collections being used as mere sub-data of e.g. nodegroups, i.e.
not being instantiated on linking, the previous check simply testing if
an object belongs to a collection was not enough enymore.

Replace it by gathering all instantiated objects (i.e. objects in
collections that are either instantiated in the viewlayer, or by a
collection instantiated by an Empty object in that viewlayer).

Pull Request: https://projects.blender.org/blender/blender/pulls/150072