Grease Pencil: Edit mode wire color
This PR adds a setting for controlling the color of Grease Pencil edit
mode wires to the theme. Because it is extremely common for artists to
use black lines, the default color should be visible against black.
A transparent grey works for this (Same as legacy Grease Pencil).

This also makes the point be drawn on top of the lines in edit mode.

Pull Request: https://projects.blender.org/blender/blender/pulls/144674