Fix #150122: VSE: Playback controls in preview use active scene context
The `playback_controls` function was checking for the `view_type` to
change the scene context for the sequencer. But this should be done
for any of the `view_type`s so the fix removes that part of the
condition.

This means that the scene syncing will show up in all views, but that's
probably better anyway and more consistent.

Pull Request: https://projects.blender.org/blender/blender/pulls/150279