Fix: UI: Activating modifier wrongly consumes event in some cases
The `OBJECT_OT_modifier_set_active` operator is registered in the
properties editor and is called even in cases where the editor is not
in the modifiers tab, just by clicking in the region.

In `edit_modifier_invoke_properties_with_hover` if no
modifier was found under the cursor, the operator would return
`OPERATOR_CANCELLED` which causes the click event to be consumed
and any other operator that might need the event is thus never invoked.

This fixes the issue by making sure that `OPERATOR_PASS_THROUGH` is also
returned so that the event can be passed on.

Pull Request: https://projects.blender.org/blender/blender/pulls/149928