UI: 3D Viewport: Add performance timing info
This adds a new overlay options to view performance timings
of the 3D viewport.  These timings are measuring the
time it takes to prepare the data for rendering on the GPU.
For now, we don't have an easy way to measure timings
on the GPU itself (e.g. the time it took to render the frame).
But since this is happening asynchronously, it's probably less
important for users. The playback frame rate is mostly affected
by the following:

* `Evaluation`: The time it took the dependency graph to
    evaluate (e.g. to get the evaluated state of the objects for rendering).
* `Synchronization`: The time it took to build/synchronize the
   data that will be uploaded to the GPU for rendering.

The `Overlays` panel is updated to accommodate for the additional
toggle for `Text` related overlays.

Note that this is not meant to give more insight into scene evaluation
timings (other than the total), it's meant for viewport rendering
related information.

In the future we might be able to get more information from the GPU
to add to the list.

Pull Request: https://projects.blender.org/blender/blender/pulls/149523