Fix #150306: eevee recompiles for minor changes in normal map node
The code was wrongly using `GPU_constant` before which implies that the value
would be inlined into the generated shader code. This resulted in changing
shader code whenever the strength input is changed, which triggered a
recompilation.

Now the node always uses `GPU_uniform`. The check for `node->runtime->original`
seems to be a relic of the past.

Pull Request: https://projects.blender.org/blender/blender/pulls/150346