Refactor: Mesh: Use strings to store active UV map
For #122398, we need to move the storage of whether a UV map is active
or default out of CustomData. For colors attributes, and in the 5.0+
file format, we do that by storing strings referencing the names of
those attributes sepoarately on the mesh. This commit switches the way
the status is stored at runtime. Most of the work has happened already
as part of #122398, this just changes the last few places and simplifies
propagation of the active status where now possible.

For BMesh, the active/default status still uses the existing CustomData
system; the change would just be far too large if I started changing
that too: all the places that use `BM_uv_map_offsets_get` would have to
change.

Like for color attributes, this change makes setting the active/default
UV map a more explicit process.

Pull Request: https://projects.blender.org/blender/blender/pulls/150112