Fix #150451: Crash when using the redo panel while using "solo" bone collection
This is a regression in 5.0, as far as I can tell this worked in 4.5
When using the redo panel for the operator `WM_OT_context_collection_boolean_set`
(Alt + W in the 3d viewport in pose mode) blender would segfault.
The issue was that the `bPoseChannel.bone` pointers were not rebuilt properly on undo/redo.

In `memfile_undo.cc` in the function `memfile_undosys_step_id_reused_cb` the pose would be
tagged for a rebuild with `ob_self->pose->flag |= POSE_RECALC;` but that rebuild didn't happen
until the call to `screen_ctx_selected_pose_bones`.
So the fix is to check for that flag, and if set, rebuild the pose.

Likely related to the change 8ebce22643f84baa35d2992bf9b5fe9a9d266723 which made
the crash visible, even though the issue is far deeper than that.

Pull Request: https://projects.blender.org/blender/blender/pulls/150513