Fix #150374: Default to opening OpenEXR files as Linear Rec.709 again
Previously it was using the working space as default, which is common behavior
in various apps. However for compatibility with existing EXR files produced by
Blender or HDRIs from polyhaven this can be suboptimal.

Now that Blender writes color interop IDs it is not important for EXR files
produced by Blender itself, and as more apps start doing the same the default
should not matter as much in the future.

Pull Request: https://projects.blender.org/blender/blender/pulls/150493