Vulkan: Graphics Pipeline Libraries
This PR will separate compiling the more reusable parts of a graphics
pipeline to improve the caching and reduce the needed compilation.
It is based on the VK_EXT_graphics_pipeline_library extension.

The benefits (in performance) at this moment are not visible and will
require more changes. This PR was done in first place as it provided
information to design an API for pipeline warm-up. (part of #148398).

The graphics pipeline is spliced into 3 libraries; vertex input, shaders
and fragment output. Shaders include all shaders to ensure that the
interfaces between the shaders will be compatible.

Building and caching the different libraries uses the same locking
mechanism we introduced for compute and (executable) graphics
pipelines.

Downside of the implementation is the over utilization of stack space.
This can be improved in the future by separating
`VkGraphicsPipelineBuilder` into multiple smaller structs (or by
using a template for state).

For the time being this PR has been disabled on ghost level for AMD
drivers due to instability. We are working with AMD to also find out
why this is the case.

Pull Request: https://projects.blender.org/blender/blender/pulls/148248