Fix #150166: Broken image when rendering image in VSE
When render size is set to percentage not equal to 100% and opengl
render is done, the image is broken. This is, because render operator
expected image size equal to scene size, but image returned by sequencer
is smaller. That happens, because final cache did not use render size
as part of its key.

Pull Request: https://projects.blender.org/blender/blender/pulls/150402