RNA: add callback to support dynamic property ui description
This PR adds a new `PropertyRNA::description_func` property similar to the
`PropertyRNA::name_func` added as part of PR #147137. This allows dynamically
getting a UI description for the property.

This is useful for displaying correct tooltips for node sockets which are added
as properties in UI panels. Without this, they use the hardcoded description
"Input value used for unconnected socket" which is not helpful. This patch shows
the correct description as specified in the node socket.

Pull Request: https://projects.blender.org/blender/blender/pulls/150403