Assets: include Blender version + platform in remote asset listing request
Set the `User-Agent` HTTP header on all Online Assets HTTP calls. This
is set to the same string that's already used by the Extensions API
calls: `Blender/5.1.0 (Linux x86_64; cycle=alpha)`. It is used for all
Online Asset HTTP calls, so also those that download the assets and
the thumbnails.

Originally this PR would put this information in the query string for
`GET` requests to the JSON files. It was chosen not to do this to
lighten the load on any caching infrastructure, as otherwise that
cache will have to store a copy of each file for each permutation of
(blender version, platform).

Pull Request: https://projects.blender.org/blender/blender/pulls/148741