Refactor: Shapekey code
No functional changes intended.

While investigating performance bottlenecks I had a look
at the shapekey code. It turns out this code is very old
and not conforming to the modern style.
This patch refactors the code by:
* Fixing comments to conform to the styleguide
* spelling out variable names instead of using abbreviations
(for the few places where I understood the point of the variable)
* moving the variable declaration closer to the place where it is actually used
* using the `LISTBASE_...` macros to iterate lists
* marking variables as const when applicable
* mark code as unreachable with an assert instead of a comment

There is more work to be done to clean this up.

Pull Request: https://projects.blender.org/blender/blender/pulls/150633