Fix: invalid duplicate face check in BM_mesh_is_valid
Correct logical error in [0] which treated faces as a set of vertices.
This isn't correct as multiple faces may share the same vertices
in a different order (with different edges).

Resolving using ordered vertex indices starting with the lowest,
using the lowest adjacent index to account for winding in either
direction.

[0]: ba83825cfaa00557d4770e957069c864a85652fa