Refactor: Un-indent shapekey code
No functional changes intended.

This patch restructures the loop of `key_evaluate_relative`
such that any conditions under which evaluation is skipped
are checked before and trigger a `continue`.
This makes it much easier to parse the flow of the code.

Also move a variable into a narrower scope.
`KeyBlock *k[4]` is only used in the `else` part
of certain functions.

Pull Request: https://projects.blender.org/blender/blender/pulls/150692