Refactor: Geometry Nodes: avoid tree references in mapping
This replaces usages of e.g. `bNode *` with the corresponding integer
identifiers. This helps to make the evaluation code less dependent on the node
tree that is used for lookup in these maps which is required for allowing
closures to be passed between objects.

Pull Request: https://projects.blender.org/blender/blender/pulls/150707