UI: Don't break playhead when drawing subframes
When scrubbing the playhead with subframes enabled, the line head would detach from the line.
This is fixed by placing the head onto the subframe as well.
To indicate that subframes are enabled, and to better show the current placement,
the playhead now shows the subframe position with 2 digits after the decimal point.

Pull Request: https://projects.blender.org/blender/blender/pulls/150578