EEVEE: Add support for Transparent BSDF through Shader To RGB
This PR adds to possibility to get the radiance behind an object using
the Shader To RGB workflow.

The result of this operation depends on the type of material.
- Dithered Material with Raytraced Transmission **off** will sample the nearest probe.
- Dithered Material with Raytraced Transmission **on** will show Dithered Materials
  with transmission **off** or fallback to nearest probe.
- Blended Material will show all Dithered Material materials or fallback to nearest probe.

This covers some workflows that were possible before 4.2 (which
were relying on the refraction BSDF).

This removes the `use_combined_lightprobe_eval` optimization which was defaulting
all Dithered Material to be in the same layer when raytracing was off.

We cannot do that anymore since this transparency feature requires clear separation
between the two material types (Raytraced Transmission on/off).
We could eventually reintroduce it by changing the way the submission is done when
there is no shader to rgb in the scene.

Pull Request: https://projects.blender.org/blender/blender/pulls/149723