Fix: UI: Wrong emboss in some widgets
The check for whether to draw the emboss was checking for outline
and inner alpha other than 0, which at first sight seemed fine.

However, this didn't account that on hover sometimes the inner opacity
is increased for highlighting (for example `menu_item`), thus drawing
the emboss even though the outline is fully transparent, making it look
like a weird gap.

Fix this by just checking for the outline's alpha, and make that check
outside "if draw_outline", since it can happen that widgets have that
disabled but will still draw emboss. It's a bit more flexible this way.

This was a mix of an oversight by me in !111980, but also back then
we didn't have the concept of active and hover menu items as two
separate things.

Pull Request: https://projects.blender.org/blender/blender/pulls/150766