Cleanup: quiet erroneous "alloc-size-larger-than" warnings with GCC
GCC 15.2.1 release builds were incorrectly warning the allocation size
to MEM_malloc_arrayN is an invalid range.

Delaying the old-to-new map creation in BM_mesh_remap works around this.
Apply the workaround as it's a minor improvement, also replace unique
strings with __func__ as it's not needed for each allocation to have a
unique name.