UI: Center Align Non-Overlapped Header Status Text
With #139507 we are center aligning the changing operator values in
overlapped header regions. This seems to be successful so we should
consider doing so for non-overlapped headers as well.  We have a
combination of header overlaps so this would make them consistent.

Pull Request: https://projects.blender.org/blender/blender/pulls/149857