Fix #150131: Selection state of edit bone root/tip not preserved
When switching modes on an armature from edit -> pose -> edit mode,
the selection states of the root and tip were lost. That was caused
by the move to storing the selection state on the pose bone itself, which
does only deal with a full bone selection.
See a09d0cfd8ccecb57248c6ed96ba25900eab82e0f

The fix is to store the root/tip state on the pose bone, even though it is
not used for the pose itself.

Pull Request: https://projects.blender.org/blender/blender/pulls/150892