Vulkan: Vertex Input Dynamic State
This PR reduces the amount of needed pipelines by making
vertex inputs of the pipeline dynamic. They will be part
of the command buffer.

This reduces the complexity and increases the effectiveness
of !150583 as the vertex inputs don't need to match at all.
The current implementation would create to many pipelines
and assumed that the driver would reuse linking strategies
which is far fetched.

Some platforms can have poorer performance, but desktop
GPUs should be OK. As a fallback is available it  can be
disabled easily.

Note: Feature is disabled for official AMD drivers due
to stability issues.

Pull Request: https://projects.blender.org/blender/blender/pulls/150951