Geometry Nodes: Register separate operator for each node tool
Reusing the same operator type for all node tools is stretching the
limits of Blender's operator design too much, especially as node tools
become more featured and less homogenous. This is especially bad for
modal node tools in the future, where customizing a modal keymap
wouldn't work without significantly increased complexity. Other things
that don't work in the current design is remembering properties from an
operator's last execution, and having proper poll functions.

This PR changes the system to register a separate operator for every
asset and local tool. A new string property on the node group controls
the operator idname. This is set by default when creating a new tool, but
should be customized when creating an asset for sharing across files.

The change is not fully backward compatible. For existing node tools
to work with 5.1, the files must be opened and saved in 5.1 as well.
If the tool's name is not unique, the automatically chosen operator
needs to be changed.

Invalid or duplicate operator names are reported in the interface
where custom idnames are configured. Though this is conceptually
a more global problem, asset creators building tools to share are
expected to choose a unique name, and local node group names
are unique within a specific file anyway.

When a property that affects the registered operator changes, or
assets are loaded, these node group operators are re-registered.

Pull Request: https://projects.blender.org/blender/blender/pulls/148763