Fix #151028: Marking 4.2 brushes as assets overrides with wrong preview
In the old brushes system it was possible to select an image to use as
brush preview. The image would be stored as image buffer in the brush.
Whenever the preview was displayed in the UI, the automatic preview
system would copy that image buffer into the preview buffer.

After brush assets were introduced, all this was redundant and the old
preview image code was removed in 4ccf435058. But the automatic preview
system would still allow brushes to render an automatic preview. This
would just fall-back rendering a material preview without any setting
up.

So when marking a brush as asset they would get a material preview
render as preview.

In fact, often the custom image was already loaded into the preview
buffer. This is now preserved (not overridden when marking as asset), so
the preview image set in the old system will show with the new system.