Fix: Vulkan: Vertex input library
When vertex input dynamic state is enabled the vertex input of the
pipeline should be ignored, however some drivers still parse it to be
valid. This PR will set a nullptr to the vertex inputs so these
drivers will not crash when building the vertex input library.

Pull Request: https://projects.blender.org/blender/blender/pulls/151291