Grease Pencil: Improve draw performance at beginning and end of stroke
This avoids re-triangulation of the entire drawing at the beginning and
at the end of drawing a new stroke as well as the update of the primitive
tools.

Currently, this only works in the (most common) case when drawing
on top instead of to the back of the drawing.

Instead of tagging the runtime caches for the triangulation
and the plane normals for re-computation, copy the caches
of all the strokes that didn't change and update the cache of
the new stroke immediately.

In a test file with 277 fill strokes and 51k points this reduced
the computation of the cache from >1800 ms to <0.1ms.

Pull Request: https://projects.blender.org/blender/blender/pulls/151191