VSE: Box Gesture in Blade Tool
This PR implements a box gesture for the blade tool as proposed in
#125096. The aim is to make it easier to quickly cut out undesired parts
of a video such as silence, automatically closing any resulting gaps in
the process.

After specifying the box size and location, all strip content within the
box is deleted. By default, future strips on the same channel are moved
backward ("rippled") to close the gap. Pressing shift before
dragging keeps the gap if desired (also configurable from tool settings).

The ripple behavior is applied to all channels that intersect the box.
This allows strips on specified channels to remain at the same relative
location.

If the box is at the start of a strip, the amount to ripple is the
distance from the left handle to the right side of the box (this
behavior is mirrored when cutting the ends of strips). If the cut
"straddles" two strips, it will remove the gaps if the option is set.

While proper ripple delete / trim tools would be able to accomplish
something similar (and are planned for future work), combining
split/cuts and ripple into one action can accelerate the editing
workflow.

Co-authored-by: John Kiril Swenson <eliphaz@noreply.localhost>
Pull Request: https://projects.blender.org/blender/blender/pulls/145999