Fix: Crash rendering subdivision in EEVEE
The issue has been introduced by !149681: the subdiv_face_offset is
no longer guaranteed to be allocated using guarded allocator (for
example, when a small buffer optimization is used), and the drawing
was using MEM_dupallocN() on it.

Pass the subdiv_face_offset as a Span, to simplify making copies in
a more C++-friendly way.

Pull Request: https://projects.blender.org/blender/blender/pulls/151298