Fix #150957: Image changes in 'Single Image' mode don't update viewport
Introduced with 2131ef0a2093cd2a19dc925b503e4dca41eddedb

With the above commit, the source of truth for determining if an engine
needed to be updated based on changes was moved to the depsgraph update
count. This had the side effect of causing image undo and painting in
the Image Editor to not update the 3D Viewport for the 'Single Image'
mode, as this special case does not have the image exposed as part of
the shader.

To fix this, check for this specific mode and tag the corresponding
object for a shading update to force the recalculation.

Further work should be done in this area to determine if this mode needs
to be maintained in this exact form, as a special override.

Pull Request: https://projects.blender.org/blender/blender/pulls/151087

Pull Request: https://projects.blender.org/blender/blender/pulls/151363