Fix: USD: Make object names more deterministic during import
It's possible for certain constructs in USD to result in
non-deterministic patterns in Blender across different imports (or the
same import if there are other identical constructs in the file). For
example, given a USD prim hierarchy at path `/root/A/A`, this can result
in either an object hierarchy of `A/A.001` or `A.001/A` in Blender.

This is mostly due to our sorting step which attempts to work around
inefficiencies when creating large amounts of objects; creating them in
order by name is faster. Since the incoming names are the same, e.g `A`
and `A`, sometimes the order is swapped.

To make things consistent now, and to alleviate problems in the future
with concurrent import, it's important to keep the order stable. We
break the tie on name by using the original USD path instead.

Pull Request: https://projects.blender.org/blender/blender/pulls/151281