Fix #151206: Crash When Switching From Normal Rotate to Transform Rotate
A refactor in [0] added a block of code for `TFM_NORMAL_ROTATION` which
bypassed  the creation of a full `TransData` and instead allocated a
`Dummy` container  with  empty data. Subsequent switching to the
transform rotate mode, calls `ElementRotation_ex` which runs
`add_v3_v3v3(td->loc, vec, center);` but since `td->loc` is null,
resulting in a crash.

Removes the code added in [0] to allow a valid `TransData` to be
created for transform rotation.

Ref !151219

[0]: 58af30f9b90d1e6c649f9a093fa8ca5732028837