UI: support search-on-key-press when adding new workspace
Previously, one had to navigate a nested menu when adding a new workspace. Now
the menu also uses search-on-key-press to make it very quick to find a specific
workspace.

This also adds the `Search...` entry at the top to be consistent with other add
menus.

Currently, the patch removes the plus icon in the header. This is mainly done
because that's not supported by `MenuType` afaik. Adding support for could be
done separately. Not having the additional icon there is consistent with all
other add menus.

Pull Request: https://projects.blender.org/blender/blender/pulls/151317