Refactor: Change `PaintStroke` to virtual `struct`
This commit refactors the `PaintStroke` struct to be exposed in
`paint_intern.hh`. All but a few of the members used for initialization
are set as private or protected, as necessary.

The existing callbacks are now `virtual` protected methods for each of
of the paint modes to override.

Ref: #149378

Pull Request: https://projects.blender.org/blender/blender/pulls/151220