Fix: Geometry Nodes: crash when capturing group input geometry in some cases
The crash requires a fairly specific setup that I noticed when working on
#150852. The setup is included here in form of a regression test.

The issue happened because the closure currently assumes that the Group Input
geometry can be part of the output because it is captured due to being a border
link. In another pass when computing the final propagate-relations (data that
can propagate from the group input to group output) it is detected that this
can't actually happen. So there was some inconsistency in the final required
reference lifetimes.

There might be a cleaner fix that doesn't require a post-processing step, but I
haven't found one yet. Might need some more investigation or refactoring later.

Pull Request: https://projects.blender.org/blender/blender/pulls/151390