Refactor: shapekey - simpler copy function
The shapekey code has a function that copies the data
of a `KeyBlock` into a float array (`copy_key`).
That function is very cryptic considering all it is doing is copying
data between float arrays.
By making the function more specific it can be simplified and a lot more readable.

The old function `copy_key` is still being called in a few places,
but to deliver the change in small steps this kept for another refactor.

Pull Request: https://projects.blender.org/blender/blender/pulls/151376