Fix: Compositor file output tests not using Vulkan
The issue was the `--gpu-backend` argument was passed to the python
script and not to Blender. This is because `blender_test` calls blender
first and then runs the test script inside it, whereas a `python_test`
calls blender with the provided python arguments.

The command `gpu.platform.backend_type_get()` was used to verify the
correct gpu backend was used.

Pull Request: https://projects.blender.org/blender/blender/pulls/151478