Fix #148046: TIMER event caused unexpected behaviour on pen tool
The issue was caused by the Pen tool modal operator not handling the
TIMER event properly. The TIMER event seems to be passed, whenever the
UI gets redrawn.

The pen tool modal code was changed so that if an unrelated event is
passed to the modal function, it will fail a event type check and not
do anything.

Pull Request: https://projects.blender.org/blender/blender/pulls/148131