Fix: Out of bounds memory access with curves of size 1
This has been in Blender since forever but only now came up
because of the test `bl_animation_shapekey` which was added in
9725e1d3860514256d908229eb50a1ab7c9ab717

In that case we create a curve with a single point and then add
shapekeys to it. This fails under ASAN with a heap-buffer-overflow
since the code is set up to start the loop by creating a pointer to
the next BezTriple assuming it exists which of course it doesn't
for curves with only 1 element.

I assume that bug never came up since adding shapekeys
to curves with only 1 point is not particularly useful.

Pull Request: https://projects.blender.org/blender/blender/pulls/151664