Sculpt: Expose simple BVH raycast function for active element retrieval
Currently, we have no simple function that allows accessing basic active
element information in paint modes that use the BVH. It is assumed that
this data is retrieved from the `SculptSession` as needed and updated on
mouse movement or as necessary.

This has a number of drawbacks:
* It is often unclear what exact data is needed for a given brush or
  tool when the corresponding `cursor_geometry_info_update` function is
  called.
* It is specifically tied to sculpt mode, performing various checks and
  updates to ensure the state of the BVH and the scene in general prior
  to doing the raycast

To simplify usage for both external callers (e.g. geometry nodes) and
internal usages that only need a limited subset of data, this commit
introduces a new API specific for this simple usecase.

Part of #151686

Pull Request: https://projects.blender.org/blender/blender/pulls/151687