Fix: USD: Bspline basis curves using 'varying' widths were incorrect
The bspline curve type was untested and had issues loading in 'varying'
curve widths (it would access invalid width values). The 'varying'
interpolation is, in general, not well supported in Blender for this
type but we can be safer and get much closer to ground-truth if we load
in the width values more carefully.

Add test coverage too.

Pull Request: https://projects.blender.org/blender/blender/pulls/151598