Vulkan: Reduce memory pressure when uploading textures
Memory pressure is a pain point of vulkan backend. There are extensions
that can help with this. One of them is VK_EXT_host_image_copy that
reduces the need of a staging buffer to upload data to an image. It
allows directly uploading from an application owned buffer and perform a
layout transition afterwards.

Using this extension allows uploading initial data to an image outside our
rendergraph. Reducing the number of nodes/processing as well.

Support for this extension is disabled in Mesa/RADV due to performance
issues. It can be activated by setting setting the `RADV_PERFTEST=hic`
environment variable. It is currently disabled as it has a performance
hit for optimal tiled images, which we use.

The VK_HOST_IMAGE_COPY_MEMCPY feature isn't used due its added
complexity to match subresource layouts.

Pull Request: https://projects.blender.org/blender/blender/pulls/149256