Refactor animdata handling in node group operators
AnimData in node group/ungroup/insert/separate operators needs to be
copied and/or moved from one tree to another. The current solution is
written in outdated C style and is unnecessarily verbose and hard to
follow.

This is a refactor of the `BKE_animdata_transfer_by_basepath` and its
use in the node group operators. C++ structures help reduce verbosity
and make it easier to verify behavior.

`BKE_animdata_transfer_by_basepath` function is split into two variants:
- `BKE_animdata_copy_by_basepath` makes a copy of animdata and keeps the
  source ID unmodified (with proper const-ness too).
- `BKE_animdata_move_by_basepath` moves the animdata from source to
  destination. Both IDs are modified and have to be mutable (same as the
  original `transfer` function).

This distinction makes it possible to avoid unnecessary updates of the
source ID when the data is only copied. It also removes the need to make
a full copy of the node group when ungrouping, which was previously
necessary to avoid destroying animdata in the original group.

Pull Request: https://projects.blender.org/blender/blender/pulls/151668