Outliner: `OUTLINER_OT_show_active`: skip tree_element if it's not in the collection
# Motivation
When executing `OUTLINER_OT_show_active` in the outliner, the operator would focus the view on the first tree_element it encountered with the matching id. This is especially useless/problematic when the active object is not in the parent collection of the corresponding tree_element. We can't toggle the selection, hide, or toggle viewport/render as those buttons are not (nor they should be) drawn for such tree_elements to avoid UI clutter. It makes traversing the tree very difficult in a complex hierarchy which is inevitable in a production environment.

# Implementation:
This PR adds a new parameter to `outliner_find_id()` called exclude_flags. Then `outliner_show_active_element_get()` calls the `outliner_find_id()` with `TE_CHILD_NOT_IN_COLLECTION` as the exclude_flags argument. This way, any tree_element that has the exclude_flags is skipped even if the id matches.

`outliner_find_id()` is also called from `parent_drop_set_parents()` but we don't want to exclude any tree_element so pass `(TreeElementFlag)0` to exclude_flags.

Pull Request: https://projects.blender.org/blender/blender/pulls/151604