Anim: avoid copy when blending layers
During animation evaluation the `blend_layer_results` function took in
two `EvaluationResult` instances, and blended them into a new result.
This has to create and allocate a new copy of `EvaluationResult`.
The more efficient way to do this is to modify the result in place
and merge the results into it.

This currently has no measurable performance delta because
animation layers are not yet exposed.

Pull Request: https://projects.blender.org/blender/blender/pulls/150592