Fix #151760: Adjusting pose asset refreshes thumbnail
When updating a pose asset, the thumbnail used to be
regenerated. That means it potentially deletes user
generated data without any way to undo this.

The fix is to remove the code that automatically
generates the preview. The user can always go
to the asset browser and use the "Generate Preview"
button there.

Ideally we would tag auto generated thumbnails, so
they can be updated automatically while user specified
thumbnails are kept. Since this patch should be backported
to 4.5 I kept the change as simple as possible.

Pull Request: https://projects.blender.org/blender/blender/pulls/151818