XR: Increase teleportation arc control points amount, cleanup use of macro
Small follow-up/tweak to the XR arc teleportation rewrite (899ce2e182).

This increases the number of control points used for generation from 12
to 24, effectively raising the length of the initial arc used for
scene raycast intersection. This fixes a small usability issue where the
arc could sometimes be too short to reach its target, when aiming
straight up or when trying to teleport down from a high platform in a
large scale scene for example.

Additionally, this moves the control points amount constant from a
global GPU header define to a static constexpr variable limited to
wm_xr_operators.cc (its only use place). The GPU header was a remainder
of the previous teleportation arc implementation (which used a custom
shader) and with this constant being the only thing left it in it, it
has now been removed.