UI: reduce size of the marker region
The marker region was quite big. When visible in the timeline of
the startup file, it blocks most of the time scrubbing area.

With this patch it is shrunk in size a bit by:
* Reducing the actual size by 6
* Aligning text with the marker icon
* reducing the delta for text offsetting

Note that I removed the background fade for these images as proposed in #148633

We still have to have quite a big margin below so the scroll bar can grow into it
This makes the timeline just big enough for the summary keys to be visible when markers are shown

Pull Request: https://projects.blender.org/blender/blender/pulls/148635