VSE: Set preview frame to opposite side when moving blade box
This PR addresses the problem that it is currently not possible to see
in the preview on what frame the opposite side of the the box is and
when adjust the position of the box there is no way of seeing in the
preview where the start of the box is moved to.

The Problem is solved by changing the preview frame to preview be other
side of the box when moving the box, that makes it possible to see
where the box is moved to.

Pull Request: https://projects.blender.org/blender/blender/pulls/151569