Fix #150392: Sculpt Texture Paint doesn't update 3D Viewport when painting on single image texture
When using experimental texture paint and setting the canvas selector
to image mode, the viewport doesn't update to show the image texture on
the object being painted.

This is because this functionality requires setting the
`image_paint_override` to the current image in the canvas selector.
Experimental texture paint and legacy texture paint use different
underlying objects to store the current canvas image, those being
`PaintModeSettings` and `ImagePaintSettings` respectively, so the fix
is just to add some code to set the `image_paint_override` using the
`PaintModeSettings` object when the user is texture painting in sculpt
mode.

Pull Request: https://projects.blender.org/blender/blender/pulls/151405