Fix #147784: Darken Starting Area That Is Closed When Docking
In 4.3 - 4.5 while dragging an area to a new location the source area
would darken to indicate that it would be closed at the end of the
operation. Unfortunately in 5.0 with #144591 I broke that while making
changes necessary for the area menu. This PR is just a one-line change
that restores that darkening effect.

Pull Request: https://projects.blender.org/blender/blender/pulls/151928