DNA: make most ListBase in DNA typed
This is a follow-up to #138480. It changes the type from `ListBase` to
`ListBaseT<...>` in almost all cases. There are very few exceptions where I
couldn't easily figure out the type now (mostly because those members are
legacy).

There are also a few changes outside of `makesdna` which were necessary for
things to compile.

This does not update the usage of the lists yet (like using the C++ iterator).
This should be done separately.

In some places I've used forward declarations. Either because the actual type is
internal to some module, or because I didn't want to move around structs as part
of this commit as that results in a lot of noise.

Pull Request: https://projects.blender.org/blender/blender/pulls/151970