Fix #151989: Crash when using the Alpha output of an image
Blender crashes when using the Alpha output of an image that has a
combined pass that is not a color. This is because alpha extraction
assumes a color pass. To fix this, we only display an alpha pass if
the combined pass is an RGBA pass.

Pull Request: https://projects.blender.org/blender/blender/pulls/152023