Fix: Vulkan: Validation Error Input Attachments
When running Blender on certain platforms (can be simulated with
`--debug-gpu-force-workarounds`) EEVEE stencil classification generates
incorrect barriers. In the command buffer the same barrier is recorded
twice. The root cause was that the subimage of the access info wasn't
filled -> two layers of the same image created a barrier for the full
image -> validation layer.

For most desktop GPU this didn't lead to an issue, but mobile GPUs are
more strict on this.

Fixed by extracting the correct subimage.

Pull Request: https://projects.blender.org/blender/blender/pulls/152034