Fix #152018: Crash if overflowing scene statistics buffer
The main issue was not accounting for the current offset (`ofs`) in a
call to `BLI_snprintf_utf8_rlen`. This would cause an overflow of the
`info` buffer if a suitably long Object name were already included in
the output.

A secondary issue is that the buffer is too small now that we allow
longer Object names. This would cause the statistics output to be
clipped unnecessarily. Increase the buffer to better account for this.

Pull Request: https://projects.blender.org/blender/blender/pulls/152025