Fix: Nodes: Allow transparency in node headers
Allow the theme to set transparency in node headers.
This was possible before !140481, which made headers use the exact
theme color but hardcoded transparency to be fully opaque.

This is not to be ported to previous LTS since it only affects
the 5.x node redesign.

See PR for details and screenshots.

Pull Request: https://projects.blender.org/blender/blender/pulls/149448